+++ compiler/z/codegen/SystemLinkage.cpp
@@ -19,9 +19,6 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-// See also S390Linkage.cpp which contains more S390 Linkage
-// implementations (primarily Private Linkage and base class).
-
 #include <algorithm>
 #include <stddef.h>
 #include <stdint.h>
@@ -70,26 +67,24 @@
 #include "z/codegen/S390Evaluator.hpp"
 #include "z/codegen/S390GenerateInstructions.hpp"
 #include "z/codegen/S390Instruction.hpp"
-#include "z/codegen/TRSystemLinkage.hpp"
+#include "z/codegen/SystemLinkage.hpp"
 
-#ifdef J9_PROJECT_SPECIFIC
-#include "env/VMJ9.h"
-#endif
-
-#define  GPREGINDEX(i)   (i-TR::RealRegister::FirstGPR)
-
-////////////////////////////////////////////////////////////////////////////////
-// TR::S390zOSSystemLinkage for J9
-////////////////////////////////////////////////////////////////////////////////
+TR::SystemLinkage::SystemLinkage(TR::CodeGenerator* cg, TR_S390LinkageConventions elc, TR_LinkageConventions lc)
+   : 
+      TR::Linkage(cg, elc,lc),
+      _GPRSaveMask(0), 
+      _FPRSaveMask(0)
+   {
+   }
 
 TR::SystemLinkage *
-TR::S390SystemLinkage::self()
+TR::SystemLinkage::self()
    {
    return static_cast<TR::SystemLinkage *>(this);
    }
 
 void
-TR::S390SystemLinkage::initS390RealRegisterLinkage()
+TR::SystemLinkage::initS390RealRegisterLinkage()
    {
    int32_t icount = 0, ret_count=0;
 
@@ -121,7 +116,7 @@ TR::S390SystemLinkage::initS390RealRegisterLinkage()
 
 // Non-Java use
 void
-TR::S390SystemLinkage::initParamOffset(TR::ResolvedMethodSymbol * method, int32_t stackIndex, List<TR::ParameterSymbol> *parameterList)
+TR::SystemLinkage::initParamOffset(TR::ResolvedMethodSymbol * method, int32_t stackIndex, List<TR::ParameterSymbol> *parameterList)
 {
    ListIterator<TR::ParameterSymbol> parameterIterator((parameterList != NULL)? parameterList : &method->getParameterList());
    parameterIterator.reset();
@@ -177,7 +172,7 @@ TR::S390SystemLinkage::initParamOffset(TR::ResolvedMethodSymbol * method, int32_
  * The operation depends on supplied opcode
  */
 TR::Instruction *
-TR::S390SystemLinkage::getputFPRs(TR::InstOpCode::Mnemonic opcode, TR::Instruction * cursor, TR::Node *node, TR::RealRegister *spReg)
+TR::SystemLinkage::getputFPRs(TR::InstOpCode::Mnemonic opcode, TR::Instruction * cursor, TR::Node *node, TR::RealRegister *spReg)
    {
    int16_t FPRSaveMask;
    int32_t offset, firstSaved, lastSaved, beginSaveOffset;
@@ -228,7 +223,7 @@ TR::S390SystemLinkage::getputFPRs(TR::InstOpCode::Mnemonic opcode, TR::Instructi
  *   instruction  sequence
  */
 TR::Instruction *
-TR::S390SystemLinkage::addImmediateToRealRegister(TR::RealRegister *targetReg, int32_t value, TR::RealRegister *tempReg, TR::Node *node, TR::Instruction *cursor, bool *checkTempNeeded)
+TR::SystemLinkage::addImmediateToRealRegister(TR::RealRegister *targetReg, int32_t value, TR::RealRegister *tempReg, TR::Node *node, TR::Instruction *cursor, bool *checkTempNeeded)
    {
    bool smallPositiveValue = (value<MAXDISP && value>=0);
    bool largeValue = (value<MIN_IMMEDIATE_VAL || value>MAX_IMMEDIATE_VAL);
@@ -269,7 +264,7 @@ TR::S390SystemLinkage::addImmediateToRealRegister(TR::RealRegister *targetReg, i
  * Reverse the bit ordering of the save mask
  */
 uint16_t
-TR::S390SystemLinkage::flipBitsRegisterSaveMask(uint16_t mask)
+TR::SystemLinkage::flipBitsRegisterSaveMask(uint16_t mask)
    {
    TR_ASSERT(NUM_S390_GPR == NUM_S390_FPR, "need special flip bits for FPR");
    uint16_t i, outputMask;
@@ -285,539 +280,11 @@ TR::S390SystemLinkage::flipBitsRegisterSaveMask(uint16_t mask)
    return outputMask;
    }
 
-
-TR::S390zOSSystemLinkage::S390zOSSystemLinkage(TR::CodeGenerator * codeGen)
-   : 
-      TR::SystemLinkage(codeGen, TR_SystemXPLink),
-      _stackPointerUpdate(NULL),
-      _ppa1(NULL),
-      _ppa2(NULL)
-   {
-   // linkage properties
-   setProperties(FirstParmAtFixedOffset);
-   setProperty(SplitLongParm);
-   setProperty(SkipGPRsForFloatParms);
-   if (TR::Compiler->target.is64Bit())
-      {
-      setProperty(NeedsWidening);
-      }
-   else
-      {
-      setProperty(FloatParmDescriptors);
-      }
-
-   setRegisterFlag(TR::RealRegister::GPR4, Preserved);
-
-   setRegisterFlag(TR::RealRegister::GPR8, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR9, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR10, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR11, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR12, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR13, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR14, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR15, Preserved);
-
-   setRegisterFlag(TR::RealRegister::FPR8, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR9, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR10, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR11, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR12, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR13, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR14, Preserved);
-   setRegisterFlag(TR::RealRegister::FPR15, Preserved);
-
-   static const bool enableVectorLinkage = codeGen->getSupportsVectorRegisters();
-
-   if (enableVectorLinkage)
-      {
-      setRegisterFlag(TR::RealRegister::VRF16, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF17, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF18, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF19, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF20, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF21, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF22, Preserved);
-      setRegisterFlag(TR::RealRegister::VRF23, Preserved);
-      }
-
-   setIntegerReturnRegister (TR::RealRegister::GPR3 );
-   setLongLowReturnRegister (TR::RealRegister::GPR3 );
-   setLongHighReturnRegister(TR::RealRegister::GPR2 );
-   setLongReturnRegister    (TR::RealRegister::GPR3 );
-   setFloatReturnRegister   (TR::RealRegister::FPR0 );
-   setDoubleReturnRegister  (TR::RealRegister::FPR0 );
-   setLongDoubleReturnRegister0  (TR::RealRegister::FPR0 );
-   setLongDoubleReturnRegister2  (TR::RealRegister::FPR2 );
-   setLongDoubleReturnRegister4  (TR::RealRegister::FPR4 );
-   setLongDoubleReturnRegister6  (TR::RealRegister::FPR6 );
-   if (enableVectorLinkage) setVectorReturnRegister  (TR::RealRegister::VRF24);
-   setStackPointerRegister  (TR::RealRegister::GPR4 );
-   setNormalStackPointerRegister  (TR::RealRegister::GPR4 );
-   setAlternateStackPointerRegister  (TR::RealRegister::GPR9 );
-   setEntryPointRegister    (TR::RealRegister::GPR6);
-   setLitPoolRegister       (TR::RealRegister::GPR8 );
-   setReturnAddressRegister (TR::RealRegister::GPR7 );
-
-   setEnvironmentPointerRegister (TR::RealRegister::GPR5  );
-   setCAAPointerRegister   (TR::RealRegister::GPR12  );
-   setDebugHooksRegister(TR::Compiler->target.is64Bit() ? TR::RealRegister::GPR10 : TR::RealRegister::GPR12);
-
-   setIntegerArgumentRegister(0, TR::RealRegister::GPR1);
-   setIntegerArgumentRegister(1, TR::RealRegister::GPR2);
-   setIntegerArgumentRegister(2, TR::RealRegister::GPR3);
-   setNumIntegerArgumentRegisters(3);
-
-   setFloatArgumentRegister(0, TR::RealRegister::FPR0);
-   setFloatArgumentRegister(1, TR::RealRegister::FPR2);
-   setFloatArgumentRegister(2, TR::RealRegister::FPR4);
-   setFloatArgumentRegister(3, TR::RealRegister::FPR6);
-   setNumFloatArgumentRegisters(4);
-
-   if (enableVectorLinkage)
-      {
-      int vecIndex = 0;
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF25);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF26);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF27);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF28);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF29);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF30);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF31);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF24);
-      setNumVectorArgumentRegisters(vecIndex);
-      }
-
-   setOffsetToFirstParm   ((TR::Compiler->target.is64Bit()) ? 2176 : 2112);
-   setOffsetToRegSaveArea   (2048);
-   setNumberOfDependencyGPRegisters(32);
-   setFrameType(TR_XPLinkUnknownFrame);
-   setLargestOutgoingArgumentAreaSize(0);
-   setOffsetToLongDispSlot(0);
-
-   setPrologueInfoCalculated(false);
-   setGuardPageSize   ((TR::Compiler->target.is64Bit()) ? 1024*1024 : 4096);
-   setAlternateStackPointerAdjust(0);
-   setSaveBackChain(false); // default: no forced stack pointer save by prologue
-   }
-
-
-/**
- * Non-Java use.
- * Calculate XPLink call descriptor for given method body (excl. entry offset).
- * This value is present in PPA1 block and also used for determining register
- * linkage for incoming floating point parameters.
- */
-uint32_t
-TR::S390zOSSystemLinkage::calculateInterfaceMappingFlags(TR::ResolvedMethodSymbol *method)
-   {
-   uint32_t callDescValue;
-
-   if (TR::Compiler->target.is64Bit())
-      {
-      return 0;
-      }
-
-   // Linkage field  (0 means XPLink)
-   // Bits 0-2 inclusive
-   //
-   callDescValue = 0x00000000;
-
-   //
-   // Return value adjust field
-   // Bits 3-7 inclusive
-   //
-
-   TR::DataType dataType = TR::NoType;
-   int32_t aggregateLength = 0;
-      TR_ASSERT( 0, "dont know how to get datatype of non-WCode method");
-   uint32_t rva = calculateReturnValueAdjustFlag(dataType, aggregateLength);
-   callDescValue |= rva << 24;
-
-   //
-   // Float parameter description fields
-   // Bits 8-31 inclusive
-   //
-
-   uint32_t parmDescriptorFields = 0;
-
-   ListIterator<TR::ParameterSymbol> parameterIterator(&method->getParameterList());
-   parameterIterator.reset();
-   TR::ParameterSymbol * parmCursor = parameterIterator.getFirst();
-   TR::Symbol *funcSymbol = method;
-
-   int32_t parmCount = 1;
-
-   int32_t floatParmNum = 0;
-   uint32_t parmAreaOffset = 0;
-   uint32_t lastFloatParmAreaOffset= 0;
-
-   bool done = false;
-   while ((parmCursor != NULL) && !done)
-      {
-      TR::Symbol *parmSymbol = parmCursor;
-      TR::DataType dataType = parmSymbol->getDataType();
-      int32_t argSize = parmSymbol->getSize();
-
-      done = updateFloatParmDescriptorFlags(&parmDescriptorFields, funcSymbol, parmCount, argSize, dataType, &floatParmNum, &lastFloatParmAreaOffset, &parmAreaOffset);
-
-      parmCursor = parameterIterator.getNext();
-      parmCount++;
-      }
-
-   callDescValue |= parmDescriptorFields;
-   return callDescValue;
-   }
-
-/**
- * XPLink utility
- * Calculate "Return value adjust" component of XPLink call descriptor
- */
-uint32_t
-TR::S390zOSSystemLinkage::calculateReturnValueAdjustFlag(TR::DataType dataType, int32_t aggregateLength)
-   {
-   // 5 bit values for "return value adjust" field of XPLink descriptor
-   #define XPLINK_RVA_RETURN_VOID_OR_UNUSED    0x00
-   #define XPLINK_RVA_RETURN_INT32_OR_LESS     0x01
-   #define XPLINK_RVA_RETURN_INT64             0x02
-   #define XPLINK_RVA_RETURN_FAR_POINTER       0x04
-   #define XPLINK_RVA_RETURN_FLOAT4            0x08
-   #define XPLINK_RVA_RETURN_FLOAT8            0x09
-   #define XPLINK_RVA_RETURN_FLOAT16           0x0A
-   #define XPLINK_RVA_RETURN_COMPLEX4          0x0C
-   #define XPLINK_RVA_RETURN_COMPLEX8          0x0D
-   #define XPLINK_RVA_RETURN_COMPLEX16         0x0E
-   #define XPLINK_RVA_RETURN_AGGREGATE         0x10 // lower bits have meaning
-
-   uint32_t rva = 0;
-
-   switch (dataType)
-      {
-      case TR::NoType:
-           rva = XPLINK_RVA_RETURN_VOID_OR_UNUSED;
-           break;
-      case TR::Int8:
-      case TR::Int16:
-      case TR::Int32:
-      case TR::Address:
-           rva = XPLINK_RVA_RETURN_INT32_OR_LESS;
-           break;
-      case TR::Int64:
-           rva = XPLINK_RVA_RETURN_INT64;
-           break;
-      case TR::Float:
-#ifdef J9_PROJECT_SPECIFIC
-      case TR::DecimalFloat:
-#endif
-           rva = XPLINK_RVA_RETURN_FLOAT4;
-           break;
-      case TR::Double:
-#ifdef J9_PROJECT_SPECIFIC
-      case TR::DecimalDouble:
-#endif
-           rva = XPLINK_RVA_RETURN_FLOAT8;
-           break;
-#ifdef J9_PROJECT_SPECIFIC
-      case TR::DecimalLongDouble:
-           rva = XPLINK_RVA_RETURN_FLOAT16;
-           break;
-      case TR::PackedDecimal:
-      case TR::ZonedDecimal:
-      case TR::ZonedDecimalSignLeadingEmbedded:
-      case TR::ZonedDecimalSignLeadingSeparate:
-      case TR::ZonedDecimalSignTrailingSeparate:
-      case TR::UnicodeDecimal:
-      case TR::UnicodeDecimalSignLeading:
-      case TR::UnicodeDecimalSignTrailing:
-#endif
-      case TR::Aggregate:
-           TR_ASSERT( aggregateLength != 0, "aggregate length is zero");
-           rva = XPLINK_RVA_RETURN_AGGREGATE;
-           if (isAggregateReturnedInIntRegisters(aggregateLength))
-              {
-              rva = rva + aggregateLength;
-              }
-           break;
-      default:
-           TR_ASSERT( 0, "unknown datatype for parm descriptor calculation");
-           break;
-      }
-
-   return rva;
-   }
-
-/**
- * XPLink utility
- * Utility return to help float parameter fields of XPLink call descriptor
- * This routine is called once for each "parameter" unless the method returns
- * false - meaning float parameter fields of the call descriptor are now fully
- * determined.
- */
-bool
-TR::S390zOSSystemLinkage::updateFloatParmDescriptorFlags(uint32_t *parmDescriptorFields, TR::Symbol *funcSymbol, int32_t parmCount, int32_t argSize, TR::DataType dataType, int32_t *floatParmNum, uint32_t *lastFloatParmAreaOffset, uint32_t *parmAreaOffset)
-   {
-   uint32_t gprSize = cg()->machine()->getGPRSize();
-
-   // Note: complex type is attempted to be handled although other code needs
-   // to change in 390 codegen to support complex
-   //
-   // PERFORMANCE TODO: it is desirable to use the defined "parameter count" of
-   // the function symbol to help determine if we have an unprototyped argument
-   // of a call (site) to a vararg function.  Currently we overcompensate for
-   // outgoing float parms to vararg functions and always shadow in FPR and
-   // and stack/gprs as with an unprotoyped call - see pushArg(). Precise
-   // information can help remove such compensation. Changes to fix this would
-   // involve: this function, pushArg() and buildArgs().
-
-   int32_t numFPRsNeeded = 0;
-   switch (dataType)
-      {
-      case TR::Float:
-      case TR::Double:
-#ifdef J9_PROJECT_SPECIFIC
-      case TR::DecimalFloat:
-      case TR::DecimalDouble:
-#endif
-         numFPRsNeeded = 1;
-         break;
-#ifdef J9_PROJECT_SPECIFIC
-      case TR::DecimalLongDouble:
-         break;
-#endif
-      }
-
-   if (numFPRsNeeded != 0)
-      {
-      uint32_t unitSize = argSize / numFPRsNeeded;
-      uint32_t wordsToPreviousParm = ((*parmAreaOffset) - (*lastFloatParmAreaOffset))/gprSize;
-      if (wordsToPreviousParm > 0xF)
-         { // to big for descriptor. Will pass in stack
-         return true; // done
-         }
-      uint32_t val = wordsToPreviousParm + ((unitSize == 4) ? 0x10 : 0x20);
-
-      (*parmDescriptorFields) |= shiftFloatParmDescriptorFlag(val, (*floatParmNum));
-
-      (*floatParmNum)++;
-      if ((*floatParmNum) >= getNumFloatArgumentRegisters())
-         {
-         return true; // done
-         }
-
-      if (numFPRsNeeded == 2)
-         {
-         (*parmDescriptorFields) |= shiftFloatParmDescriptorFlag(0x33, (*floatParmNum));  // 0x33 == bits 110001
-         (*floatParmNum)++;
-         if ((*floatParmNum) >= getNumFloatArgumentRegisters())
-            {
-            return true; // done
-            }
-         }
-      }
-   (*parmAreaOffset) += argSize < gprSize ? gprSize : argSize;
-   if (numFPRsNeeded != 0)
-      (*lastFloatParmAreaOffset) = (*parmAreaOffset);
-
-   return false;
-   }
-
-TR::Register *
-TR::S390zOSSystemLinkage::loadEnvironmentIntoRegister(TR::Node * callNode, TR::RegisterDependencyConditions * deps, bool loadIntoEnvironmentRegister)
-   {
-   // Important note: this should only be called in pre-prologue phase for routines we know are
-   // not leaf routines.
-
-   TR::Register *envReg;
-   if (loadIntoEnvironmentRegister)
-      {
-      TR::Register * systemEnvironmentRegister  = deps->searchPostConditionRegister(getEnvironmentPointerRegister()); // gpr 5
-      envReg = systemEnvironmentRegister;
-      }
-   else
-      {
-      envReg = cg()->allocateRegister();
-      }
-
-
-   return envReg;
-   }
-
-/**
- * WCode use
- */
-int32_t
-TR::S390zOSSystemLinkage::getOutgoingParameterBlockSize()
-   {
-   //
-   // Calculate size of outgoing argument area
-   //
-   int32_t argAreaSize = getLargestOutgoingArgumentAreaSize();
-   if (argAreaSize)
-      {
-      // xplink spec has minimal size
-      // we assume yes to be safe
-      int32_t minimalArgAreaSize = ((TR::Compiler->target.is64Bit()) ? 32 : 16);
-      argAreaSize = (argAreaSize < minimalArgAreaSize) ? minimalArgAreaSize : argAreaSize;
-      }
-   return argAreaSize;
-   }
-
-////////////////////////////////////////////////////////////////////////////////
-// TR::S390zLinuxSystemLinkage for J9
-////////////////////////////////////////////////////////////////////////////////
-TR::S390zLinuxSystemLinkage::S390zLinuxSystemLinkage(TR::CodeGenerator * codeGen)
-   : TR::SystemLinkage(codeGen,TR_SystemLinux)
-   {
-   // linkage properties
-   setProperties(FirstParmAtFixedOffset);
-   if (TR::Compiler->target.is64Bit())
-      {
-      setProperty(NeedsWidening);
-      setProperty(PadFloatParms);
-      }
-
-   setRegisterFlag(TR::RealRegister::GPR6, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR6, IntegerArgumentAddToPost);        // I only need to do this for r6 because r6 is the only preserved linkage register on zlinux.
-   setRegisterFlag(TR::RealRegister::GPR7, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR8, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR9, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR10, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR11, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR12, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR13, Preserved);
-   setRegisterFlag(TR::RealRegister::GPR14, Preserved);      // Setting the return address register as preserved so that way we do not always have to emit stores in prologues / epilogues
-                                                                // This possibly assumes that we don't shuffle the return address register around elsewhere
-   setRegisterFlag(TR::RealRegister::GPR15, Preserved);
-
-   // all FP registers are not preserved for zTPF
-      {
-      if (TR::Compiler->target.is64Bit())
-         {
-         setRegisterFlag(TR::RealRegister::FPR8, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR9, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR10, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR11, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR12, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR13, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR14, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR15, Preserved);
-         }
-      else
-         {
-         setRegisterFlag(TR::RealRegister::FPR4, Preserved);
-         setRegisterFlag(TR::RealRegister::FPR6, Preserved);
-         }
-      }
-
-   setIntegerArgumentRegister(0, TR::RealRegister::GPR2);
-   setIntegerArgumentRegister(1, TR::RealRegister::GPR3);
-
-   if (!comp()->getCurrentMethod()->isJ9())
-      {
-      setProperty(AggregatesPassedInParmRegs);
-      setProperty(AggregatesPassedOnParmStack);
-      // This property could be controlled by wcode INFO gcclongdouble
-      setProperty(LongDoubleReturnedOnStorage);
-      setProperty(LongDoublePassedOnStorage);
-      // This property could be controlled by wcode INFO gcccomplex
-      setProperty(ComplexReturnedOnStorage);
-      setProperty(ComplexPassedOnStorage);
-      setIntegerArgumentRegister(2, TR::RealRegister::GPR4);
-      setIntegerArgumentRegister(3, TR::RealRegister::GPR5);
-      setIntegerArgumentRegister(4, TR::RealRegister::GPR6);
-      }
-   else
-      {
-      // For JIT to native transition, We will set up param4 and parm4 in GPR8 and GPR9
-      // during param evaluation and will move them to GPR5 and GPR6 just before going
-      // into native code. This is needed for correct evaluation of parms as GPR5 is
-      // java stack pointer and GPR6 may be Literal pool pointer
-      setIntegerArgumentRegister(2, TR::RealRegister::GPR8);
-      setIntegerArgumentRegister(3, TR::RealRegister::GPR9);
-      setIntegerArgumentRegister(4, TR::RealRegister::GPR10);
-      }
-
-   setNumIntegerArgumentRegisters(5);
-
-   setFloatArgumentRegister(0, TR::RealRegister::FPR0);
-   setFloatArgumentRegister(1, TR::RealRegister::FPR2);
-   setNumFloatArgumentRegisters(2);
-
-   static const bool enableVectorLinkage = codeGen->getSupportsVectorRegisters();
-
-   if (enableVectorLinkage)
-      {
-      int vecIndex = 0;
-      // The vector registers used for parameter passing are allocated in the following order to enable
-      // future exension which might operate on vector register pairs.
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF24);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF26);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF28);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF30);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF25);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF27);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF29);
-      setVectorArgumentRegister(vecIndex++, TR::RealRegister::VRF31);
-      setNumVectorArgumentRegisters(vecIndex);
-      }
-
-   if (TR::Compiler->target.is64Bit())
-      {
-      setFloatArgumentRegister(2, TR::RealRegister::FPR4);
-      setFloatArgumentRegister(3, TR::RealRegister::FPR6);
-      setNumFloatArgumentRegisters(4);
-      }
-
-   setIntegerReturnRegister (TR::RealRegister::GPR2 );
-   setLongLowReturnRegister (TR::RealRegister::GPR3 );
-   setLongHighReturnRegister(TR::RealRegister::GPR2 );
-   setLongReturnRegister    (TR::RealRegister::GPR2 );
-   setFloatReturnRegister   (TR::RealRegister::FPR0 );
-   setDoubleReturnRegister  (TR::RealRegister::FPR0 );
-   setLongDoubleReturnRegister0  (TR::RealRegister::FPR0 );
-   setLongDoubleReturnRegister2  (TR::RealRegister::FPR2 );
-   setLongDoubleReturnRegister4  (TR::RealRegister::FPR4 );
-   setLongDoubleReturnRegister6  (TR::RealRegister::FPR6 );
-   if (enableVectorLinkage) setVectorReturnRegister  (TR::RealRegister::VRF24 );
-   setStackPointerRegister  (TR::RealRegister::GPR15 );
-   setNormalStackPointerRegister  (TR::RealRegister::GPR15 );
-   setAlternateStackPointerRegister  (TR::RealRegister::GPR11 );
-   setEntryPointRegister    (TR::RealRegister::GPR14 );
-   setLitPoolRegister       (TR::RealRegister::GPR13 );
-   setReturnAddressRegister (TR::RealRegister::GPR14 );
-
-   setGOTPointerRegister   (TR::RealRegister::GPR12  );
-
-#if defined(OMRZTPF)
-/*
- * z/TPF os is a 64 bit target
- */
-   setOffsetToRegSaveArea     (16);    // z/TPF is a 64 bit target only
-   setGPRSaveAreaBeginOffset  (16);    //x'10'  see ICST_R2 in tpf/icstk.h
-   setGPRSaveAreaEndOffset    (160);   //x'a0'  see ICST_PAT in tpf/icstk.h
-   setOffsetToFirstParm       (448);   //x'1c0' see ICST_PAR in tpf/icstk.h
-#else
-   setOffsetToRegSaveArea   ((TR::Compiler->target.is64Bit()) ? 16 : 8);
-   setGPRSaveAreaBeginOffset  ((TR::Compiler->target.is64Bit()) ? 48 : 24);
-   setGPRSaveAreaEndOffset  ((TR::Compiler->target.is64Bit()) ? 128 : 64 );
-   setOffsetToFirstParm   ((TR::Compiler->target.is64Bit()) ? 160 : 96);
-#endif /* defined(OMRZTPF) */
-   setOffsetToFirstLocal  (0);
-   setOffsetToLongDispSlot(TR::Compiler->target.is64Bit() ? 8 : 4);
-   setOutgoingParmAreaBeginOffset(getOffsetToFirstParm());
-   setOutgoingParmAreaEndOffset(0);
-   setStackFrameSize(0);
-   setVarArgOffsetInParmArea(getOutgoingParmAreaBeginOffset());
-   setVarArgRegSaveAreaOffset(0);
-
-   setNumberOfDependencyGPRegisters(32);
-   setLargestOutgoingArgumentAreaSize(0);
-   }
-
 /**
  * Front-end customization of callNativeFunction
  */
 void
-TR::S390SystemLinkage::generateInstructionsForCall(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
+TR::SystemLinkage::generateInstructionsForCall(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
       TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
       TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint)
    {
@@ -829,447 +296,15 @@ TR::S390SystemLinkage::generateInstructionsForCall(TR::Node * callNode, TR::Regi
  * @return return value will be return value from system routine copied to private linkage return reg
  */
 TR::Register *
-TR::S390SystemLinkage::callNativeFunction(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
+TR::SystemLinkage::callNativeFunction(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
       TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
       TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint)
    {
    return 0;
    }
 
-/**
- * TR::S390zOSSystemLinkage::callNativeFunction - call System routine
- * return value will be return value from system routine copied to private linkage return reg
- */
-TR::Register *
-TR::S390zOSSystemLinkage::callNativeFunction(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
-      TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
-      TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint)
-   {
-
-   /*****************************/
-   /***Front-end customization***/
-   /*****************************/
-   generateInstructionsForCall(callNode, deps, targetAddress, methodAddressReg,
-         javaLitOffsetReg, returnFromJNICallLabel, jniCallDataSnippet, isJNIGCPoint);
-
-   TR::CodeGenerator * codeGen = cg();
-
-   TR::Register * retReg = NULL;
-   TR::Register * returnRegister = NULL;
-   // set dependency on return register
-   TR::Register * lowReg = NULL, * highReg = NULL;
-
-   TR::Register * Real_highReg = NULL, * Real_lowReg = NULL, * Img_highReg = NULL, * Img_lowReg = NULL, * Real = NULL, * Imaginary = NULL;
-   switch (callNode->getOpCodeValue())
-      {
-      case TR::acall:
-      case TR::acalli:
-         retReg = deps->searchPostConditionRegister(getIntegerReturnRegister());
-         returnRegister = deps->searchPostConditionRegister(getIntegerReturnRegister());
-         if(TR::Compiler->target.is64Bit() && returnRegister && !returnRegister->is64BitReg())
-            {
-            returnRegister->setIs64BitReg(true);    //in 64bit target, force return 64bit address register,
-            }                                           //until it's known that dispatch functions can return non-64bit addresses.
-         break;
-      case TR::icall:
-      case TR::iucall:
-      case TR::icalli:
-      case TR::iucalli:
-         retReg = deps->searchPostConditionRegister(getIntegerReturnRegister());
-         returnRegister = deps->searchPostConditionRegister(getIntegerReturnRegister());
-         break;
-      case TR::lcalli:
-      case TR::lucalli:
-      case TR::lcall:
-      case TR::lucall:
-         {
-         if (TR::Compiler->target.is64Bit())
-            {
-            retReg = deps->searchPostConditionRegister(getIntegerReturnRegister());
-            returnRegister = deps->searchPostConditionRegister(getIntegerReturnRegister());
-            }
-         else
-            {
-            TR::Instruction *cursor = NULL;
-            lowReg = deps->searchPostConditionRegister(getLongLowReturnRegister());
-            highReg = deps->searchPostConditionRegister(getLongHighReturnRegister());
-
-            generateRSInstruction(codeGen, TR::InstOpCode::SLLG, callNode, highReg, highReg, 32);
-            cursor =
-               generateRRInstruction(codeGen, TR::InstOpCode::LR, callNode, highReg, lowReg);
-
-            codeGen->stopUsingRegister(lowReg);
-            retReg = highReg;
-            returnRegister = retReg;
-            }
-         }
-         break;
-      case TR::fcalli:
-      case TR::dcalli:
-      case TR::fcall:
-      case TR::dcall:
-#if defined(SUPPORT_DFP) && defined(J9_PROJECT_SPECIFIC)
-      case TR::dfcalli:
-      case TR::ddcalli:
-      case TR::dfcall:
-      case TR::ddcall:
-#endif
-         retReg = deps->searchPostConditionRegister(getFloatReturnRegister());
-         returnRegister = retReg;
-         break;
-#if defined(SUPPORT_DFP) && defined(J9_PROJECT_SPECIFIC)
-      case TR::decall:
-      case TR::decalli:
-         highReg = deps->searchPostConditionRegister(getLongDoubleReturnRegister0());
-         lowReg = deps->searchPostConditionRegister(getLongDoubleReturnRegister2());
-         retReg = codeGen->allocateFPRegisterPair(lowReg, highReg);
-         returnRegister = retReg;
-         break;
-#endif
-      case TR::calli:
-      case TR::call:
-         retReg = NULL;
-         returnRegister = retReg;
-         break;
-      default:
-         retReg = NULL;
-         returnRegister = retReg;
-         TR_ASSERT(0, "Unknown direct call Opcode %d.", callNode->getOpCodeValue());
-      }
-
-
-   if (returnRegister != retReg)
-      {
-      generateRRInstruction(codeGen, TR::InstOpCode::getLoadRegOpCode(), callNode, returnRegister, retReg);
-      }
-
-   return returnRegister;
-   }
-
-/**
- * Front-end customization of callNativeFunction
- */
 void
-TR::S390zLinuxSystemLinkage::generateInstructionsForCall(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
-      TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
-      TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint)
-   {
-   TR::CodeGenerator * codeGen = cg();
-
-   TR::RegisterDependencyConditions * postDeps = new (trHeapMemory())
-      TR::RegisterDependencyConditions(NULL, deps->getPostConditions(), 0, deps->getAddCursorForPost(), cg());
-
-   TR::Register * systemReturnAddressRegister =
-      deps->searchPostConditionRegister(getReturnAddressRegister());
-
-   //omr todo: this should be placed in the Test compiler
-   //zLinux SystemLinkage
-#if defined(JITTEST)
-      {
-
-      if (callNode->getOpCode().isIndirect())
-         {
-         TR_ASSERT(callNode->getSymbolReference()->getSymbol()->castToMethodSymbol()->isComputed(), "system linkage only supports computed indirect call for now %p\n", callNode);
-         // get the address of the function descriptor
-         TR::Register *targetReg = codeGen->evaluate(callNode->getFirstChild());
-         generateRRInstruction(codeGen, TR::InstOpCode::BASR, callNode, systemReturnAddressRegister, targetReg, deps);
-         }
-      else
-         {
-         TR::SymbolReference *callSymRef = callNode->getSymbolReference();
-         TR::Symbol *callSymbol = callSymRef->getSymbol();
-         TR::Register * fpReg = systemReturnAddressRegister;
-         TR::Instruction * callInstr = new (trHeapMemory()) TR::S390RILInstruction(TR::InstOpCode::BRASL, callNode, fpReg, callSymbol, callSymRef, codeGen);
-         callInstr->setDependencyConditions(deps);
-         }
-      }
-#endif
-
-   }
-
-/**
- * Call System routine
- * @return return value will be return reg
- */
-TR::Register *
-TR::S390zLinuxSystemLinkage::callNativeFunction(TR::Node * callNode,
-   TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
-   TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
-   TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint)
-   {
-   /********************************/
-   /***Generate call instructions***/
-   /********************************/
-   generateInstructionsForCall(callNode, deps, targetAddress, methodAddressReg,
-         javaLitOffsetReg, returnFromJNICallLabel, jniCallDataSnippet,isJNIGCPoint);
-
-   TR::Register * returnRegister = NULL;
-   // set dependency on return register
-   TR::Register * lowReg = NULL, * highReg = NULL;
-   TR::Register * Real_highReg = NULL, * Real_lowReg = NULL, * Img_highReg = NULL, * Img_lowReg = NULL, * Real = NULL, * Imaginary = NULL;
-   switch (callNode->getOpCodeValue())
-      {
-      case TR::icall:
-      case TR::icalli:
-      case TR::iucall:
-      case TR::iucalli:
-      case TR::acall:
-      case TR::acalli:
-         returnRegister = deps->searchPostConditionRegister(getIntegerReturnRegister());
-         break;
-      case TR::lcall:
-      case TR::lcalli:
-      case TR::lucall:
-      case TR::lucalli:
-         {
-         if (TR::Compiler->target.is64Bit())
-            {
-            returnRegister = deps->searchPostConditionRegister(getIntegerReturnRegister());
-            }
-         else
-            {
-            lowReg = deps->searchPostConditionRegister(getLongLowReturnRegister());
-            highReg = deps->searchPostConditionRegister(getLongHighReturnRegister());
-
-            generateRSInstruction(cg(), TR::InstOpCode::SLLG, callNode, highReg, highReg, 32);
-            generateRRInstruction(cg(), TR::InstOpCode::LR, callNode, highReg, lowReg);
-
-            cg()->stopUsingRegister(lowReg);
-            returnRegister = highReg;
-            }
-         }
-         break;
-      case TR::fcall:
-      case TR::fcalli:
-      case TR::dcall:
-      case TR::dcalli:
-#if defined(SUPPORT_DFP) && defined(J9_PROJECT_SPECIFIC)
-      case TR::dfcall:
-      case TR::dfcalli:
-      case TR::ddcall:
-      case TR::ddcalli:
-#endif
-         returnRegister = deps->searchPostConditionRegister(getFloatReturnRegister());
-         break;
-#if defined(SUPPORT_DFP) && defined(J9_PROJECT_SPECIFIC)
-      case TR::decall:
-      case TR::decalli:
-         highReg = deps->searchPostConditionRegister(getLongDoubleReturnRegister0());
-         lowReg = deps->searchPostConditionRegister(getLongDoubleReturnRegister2());
-         returnRegister = cg()->allocateFPRegisterPair(lowReg, highReg);
-         break;
-#endif
-      case TR::call:
-      case TR::calli:
-         returnRegister = NULL;
-         break;
-      case TR::vcall:
-      case TR::vcalli:
-         returnRegister = deps->searchPostConditionRegister(getVectorReturnRegister());
-         break;
-      default:
-         returnRegister = NULL;
-         TR_ASSERT(0, "Unknown direct call Opcode %d.", callNode->getOpCodeValue());
-      }
-
-   return returnRegister;
-   }
-
-/**
- * @todo (cleanup) should remove this method and use base class version.
- */
-void
-TR::S390zLinuxSystemLinkage::initParamOffset(TR::ResolvedMethodSymbol * method, int32_t stackIndex, List<TR::ParameterSymbol> *parameterList)
-{
-   ListIterator<TR::ParameterSymbol> parameterIterator(&method->getParameterList());
-   parameterIterator.reset();
-   TR::ParameterSymbol * parmCursor = parameterIterator.getFirst();
-   int32_t numIntegerArgs=0, numFloatArgs=0, numVectorArgs=0;
-   int32_t indexInArgRegistersArray = 0;
-   TR::RealRegister::RegNum argRegNum = TR::RealRegister::NoReg;
-   int8_t gprSize = cg()->machine()->getGPRSize();
-   int32_t parmInLocalAreaOffset = getParmOffsetInLocalArea();
-
-   setIncomingParmAreaBeginOffset(stackIndex);
-   while (parmCursor != NULL)
-      {
-      int32_t previousIndex=parmCursor->getLinkageRegisterIndex();
-      parmCursor->setLinkageRegisterIndex(-1);
-      TR::DataType originalDataType=parmCursor->getDataType();
-
-      if ( parmCursor->getDataType()==TR::Aggregate)
-         {
-         TR_ASSERT(0, "should not reach here due to folding of getDataTypeOfAggregateMember()");
-         TR::DataType dt = TR::NoType;
-         if(dt != TR::NoType) parmCursor->setDataType(dt);
-         else if (parmCursor->getSize() > 8) parmCursor->setDataType(TR::Address);
-         else if (parmCursor->getSize() == 8) parmCursor->setDataType(TR::Int64);
-         else if (parmCursor->getSize() == 4) parmCursor->setDataType(TR::Int32);
-         else if (parmCursor->getSize() == 2) parmCursor->setDataType(TR::Int16);
-         else if (parmCursor->getSize() == 1) parmCursor->setDataType(TR::Int8);
-         else parmCursor->setDataType(TR::Address);
-         }
-
-      switch (parmCursor->getDataType())
-         {
-         case TR::Int64:
-            if(TR::Compiler->target.is32Bit())
-               {
-               //make sure that we can fit the entire value in registers, not half in registers half on the stack
-               if((numIntegerArgs + 2) <= getNumIntegerArgumentRegisters())
-                  {
-                  indexInArgRegistersArray = numIntegerArgs;
-                  argRegNum = getIntegerArgumentRegister(indexInArgRegistersArray);
-                  }
-               else
-                  argRegNum = TR::RealRegister::NoReg;
-               numIntegerArgs += 2;
-               break;
-               }
-         case TR::Int8:
-         case TR::Int16:
-         case TR::Int32:
-         case TR::Address:
-            indexInArgRegistersArray = numIntegerArgs;
-            argRegNum = getIntegerArgumentRegister(indexInArgRegistersArray);
-            numIntegerArgs ++;
-            break;
-         case TR::Float:
-#ifdef J9_PROJECT_SPECIFIC
-         case TR::DecimalFloat:
-#endif
-            indexInArgRegistersArray = numFloatArgs;
-            argRegNum = getFloatArgumentRegister(indexInArgRegistersArray);
-            numFloatArgs ++;
-               if (isSkipGPRsForFloatParms())
-                  {
-                  if (numIntegerArgs < getNumIntegerArgumentRegisters())
-                     {
-                     numIntegerArgs ++;
-                     }
-                  }
-             break;
-         case TR::Double:
-#ifdef J9_PROJECT_SPECIFIC
-         case TR::DecimalDouble:
-#endif
-            indexInArgRegistersArray = numFloatArgs;
-            argRegNum = getFloatArgumentRegister(indexInArgRegistersArray);
-            numFloatArgs ++;
-               if (isSkipGPRsForFloatParms())
-                  {
-                  if (numIntegerArgs < getNumIntegerArgumentRegisters())
-                     {
-                     numIntegerArgs += (TR::Compiler->target.is64Bit()) ? 1 : 2;
-                     }
-                  }
-            break;
-         case TR::VectorInt8:
-         case TR::VectorInt16:
-         case TR::VectorInt32:
-         case TR::VectorInt64:
-         case TR::VectorDouble:
-            indexInArgRegistersArray = numVectorArgs;
-            argRegNum = getVectorArgumentRegister(indexInArgRegistersArray);
-            numVectorArgs ++;
-            break;
-         default: break;
-         }
-      if (argRegNum == TR::RealRegister::NoReg)
-         {
-         if ((isPadFloatParms() &&  parmCursor->getType().isFloatingPoint()) && (gprSize > parmCursor->getSize()))
-            parmCursor->setParameterOffset(stackIndex + gprSize - parmCursor->getSize());
-          else
-            parmCursor->setParameterOffset(stackIndex);
-         stackIndex += std::max<int32_t>(parmCursor->getSize(),cg()->machine()->getGPRSize());
-         }
-      else
-         {
-         // Check where to locate the linkage register 6 which is preserved on zLiux
-          if ((getIntegerArgumentAddToPost(argRegNum) !=0 && parmCursor->isParmHasToBeOnStack()) ||
-              parmCursor->getType().isVector())
-
-            {
-            TR_ASSERT( getParmOffsetInLocalArea() >=0, "ParmOffsetInLocalArea is negative");
-
-            // on local save area
-            parmCursor->setParameterOffset(parmInLocalAreaOffset);
-
-            if (parmCursor->getType().isVector())
-               parmInLocalAreaOffset += cg()->machine()->getVRFSize();
-            else
-               parmInLocalAreaOffset += cg()->machine()->getGPRSize();
-            }
-         else
-            parmCursor->setParameterOffset(getStackFrameSize() + getRegisterSaveOffset(REGNUM(argRegNum)));
-
-         parmCursor->setLinkageRegisterIndex(indexInArgRegistersArray);
-         }
-
-      if ( isSmallIntParmsAlignedRight() && parmCursor->getType().isIntegral() && (gprSize > parmCursor->getSize())) // WCode motivated
-         {
-         // linkage related
-         parmCursor->setParameterOffset(parmCursor->getParameterOffset() + gprSize - parmCursor->getSize());
-         }
-
-      if (TR::Compiler->target.is64Bit() && parmCursor->getType().isAddress() && (parmCursor->getSize()==4))
-         {
-         // This is a 31-bit pointer parameter. It's real location is +4 bytes from the start of the
-         // 8-byte parm slot. Since ptr31 parms are passed as 64-bit values, the prologue code has the
-         // spill the full 8-bytes into the parm slot. However, the rest of the code will load/store
-         // this parm as a 4-byte value, hence we have to add +4 here.
-         parmCursor->setParameterOffset(parmCursor->getParameterOffset() + 4);
-         }
-
-      parmCursor = parameterIterator.getNext();
-      }
-
-   setVarArgOffsetInParmArea(stackIndex);
-   setIncomingParmAreaEndOffset(stackIndex);
-   setNumUsedArgumentGPRs(numIntegerArgs);
-   setNumUsedArgumentFPRs(numFloatArgs);
-   setNumUsedArgumentVFRs(numVectorArgs);
-}
-
-#if 0
-void
-TR::S390zLinuxSystemLinkage::mapSingleAutomatic(TR::AutomaticSymbol * p, uint32_t & stackIndex)
-   {
-   size_t align = 1;
-   size_t size = p->getSize();
-
-   if ((size & (size - 1)) == 0  && size <= 8) // if size is power of 2 and small
-      {
-      align = size;
-      }
-   else if (size > 8)
-      {
-      align = 8;
-      }
-
-   align = (align == 0) ? 1 : align;
-   if (true)
-      {
-      // we map in a backwards fashion
-      //TODO: not sure what to do here
-      intptr_t roundup = align - 1;
-      stackIndex = (stackIndex-size) & (~roundup);
-      p->setOffset(stackIndex);
-      }
-   else
-      {
-      // we map in a backwards fashion
-      intptr_t roundup = align - 1;
-      stackIndex = (stackIndex - size) & (~roundup);
-      p->setOffset(stackIndex);
-      // we have some stacks that are getting very close
-      TR_ASSERT( (-stackIndex + size) >= (-stackIndex), "Stack index growing too big");
-      }
-   }
-#endif
-
-void
-TR::S390SystemLinkage::mapStack(TR::ResolvedMethodSymbol * method)
+TR::SystemLinkage::mapStack(TR::ResolvedMethodSymbol * method)
   	{
 		{
      	mapStack(method,0);
@@ -1282,7 +317,7 @@ TR::S390SystemLinkage::mapStack(TR::ResolvedMethodSymbol * method)
  * The offsets will be fixed up (at the end /in createPrologue) when the stackFrameSize is known
  */
 void
-TR::S390SystemLinkage::mapStack(TR::ResolvedMethodSymbol * method, uint32_t stackIndex)
+TR::SystemLinkage::mapStack(TR::ResolvedMethodSymbol * method, uint32_t stackIndex)
    {
    int32_t i;
 
@@ -1392,7 +427,7 @@ TR::S390SystemLinkage::mapStack(TR::ResolvedMethodSymbol * method, uint32_t stac
    setLocalsAreaEndOffset(stackIndex);
    }
 
-void TR::S390SystemLinkage::mapSingleAutomatic(TR::AutomaticSymbol * p, uint32_t & stackIndex)
+void TR::SystemLinkage::mapSingleAutomatic(TR::AutomaticSymbol * p, uint32_t & stackIndex)
    {
    size_t align = 1;
    size_t size = p->getSize();
@@ -1414,14 +449,14 @@ void TR::S390SystemLinkage::mapSingleAutomatic(TR::AutomaticSymbol * p, uint32_t
    p->setOffset(stackIndex);
    }
 
-bool TR::S390SystemLinkage::hasToBeOnStack(TR::ParameterSymbol * parm)
+bool TR::SystemLinkage::hasToBeOnStack(TR::ParameterSymbol * parm)
    {
    return parm->getAllocatedIndex() >=  0 &&  parm->isParmHasToBeOnStack();
    }
 
 
 TR::Instruction*
-TR::S390SystemLinkage::saveGPRsInPrologue2(TR::Instruction * cursor)
+TR::SystemLinkage::saveGPRsInPrologue2(TR::Instruction * cursor)
    {
    int16_t GPRSaveMask = 0;
    int32_t offset = 0;
@@ -1503,7 +538,7 @@ TR::S390SystemLinkage::saveGPRsInPrologue2(TR::Instruction * cursor)
    }
 
 TR::Instruction*
-TR::S390SystemLinkage::saveGPRsInPrologue(TR::Instruction * cursor)
+TR::SystemLinkage::saveGPRsInPrologue(TR::Instruction * cursor)
    {
    int16_t GPRSaveMask = 0;
    bool hasSavedSP = false;
@@ -1584,102 +619,10 @@ TR::S390SystemLinkage::saveGPRsInPrologue(TR::Instruction * cursor)
    return cursor;
    }
 
-TR::Linkage::FrameType
-TR::S390zLinuxSystemLinkage::checkLeafRoutine(int32_t stackFrameSize, TR::Instruction **callInstruction)
-   {
-   FrameType ft = standardFrame;
-
-
-   if(comp()->getOption(TR_DisableLeafRoutineDetection))
-      {
-      if(comp()->getOption(TR_TraceCG))
-         traceMsg(comp(),"Leaf Routine Detection Disabled\n");
-      return ft;
-      }
-
-   if(isForceSaveIncomingParameters())
-      {
-      if(comp()->getOption(TR_TraceCG))
-         traceMsg(comp(),"Force Save Incoming Parameters enabled\n");
-      return ft;
-      }
-
-   int32_t zeroArgsAndLocalsStackFrameSize = (((getOffsetToFirstParm() + 7)>>3)<<3);
-
-   if(comp()->getOption(TR_TraceCG))
-      traceMsg(comp(), "stackFrameSize = %d  default 0 arg and 0 local stack frame size = %d\n",stackFrameSize,zeroArgsAndLocalsStackFrameSize);
-
-   // a) frame size < 256 and b) no alloca()calls
-   if (stackFrameSize != zeroArgsAndLocalsStackFrameSize)
-      {
-      if(comp()->getOption(TR_TraceCG))
-         traceMsg(comp(), "LeafRoutine: stackFrameSize != zeroArgsAndLocalStackFrameSize\n");
-      return ft;
-      }
-
-   // b) not var arg
-   TR::MethodSymbol *methodSymbol = comp()->getMethodSymbol();
-
-
-   // c) ensure stack pointer isn't used.
-   TR::RealRegister * spReg = getStackPointerRealRegister();
-
-   if (spReg->isUsedInMemRef())       //don't need to check gprmask as these are saved in caller stack.
-      {
-      if(comp()->getOption(TR_TraceCG))
-         traceMsg(comp(), "LeafRoutine: stack pointer isUsedInMemRef = %d is set to 1\n",spReg->isUsedInMemRef());
-
-      return ft;
-      }
-
-   return ft;
-   }
-//returns true if there are no preserved registers used in the method
-bool OMR::Z::Linkage::checkPreservedRegisterUsage(bool *regs, int32_t regsSize)
-   {
-   int32_t i;
-   bool retval = true;
-
-   if(self()->comp()->getOption(TR_TraceCG))
-      traceMsg(self()->comp(), "LeafRoutine: Checking Register Usage.  Preserved Registers that are assigned:");
-
-   for( i = TR::RealRegister::FirstGPR; i <= TR::RealRegister::LastGPR-2; ++i )      //special case for r14, the return value
-      {
-      //TR_ASSERT(regs[i] == (getRealRegister(REGNUM(i)))->getHasBeenAssignedInMethod(), "Mismatch between register assigned information!!\n");
-
-      //traceMsg(comp(), "regs[%d] = %d getHasBeenAssignedInMethod = %d\n",i,regs[i],(getRealRegister(REGNUM(i)))->getHasBeenAssignedInMethod());
-
-      if (self()->getPreserved(REGNUM(i)) && (self()->getRealRegister(REGNUM(i)))->getHasBeenAssignedInMethod())
-         {
-         if(self()->comp()->getOption(TR_TraceCG))
-            traceMsg(self()->comp(), "\t%d",i-TR::RealRegister::FirstGPR);
-
-         retval = false;
-         //return false;        // fast path, using retval for debugging purposes
-         }
-      }
-
-   //check r14 usage -- don't check if it has been assigned because it could be used in a call. the regs array passed in will take this in to account.
-   if(regs[TR::RealRegister::GPR14])
-      {
-      if(self()->comp()->getOption(TR_TraceCG))
-         traceMsg(self()->comp(), "\t%d (r14)",TR::RealRegister::GPR14-TR::RealRegister::FirstGPR);
-
-      retval = false;
-      // return false;
-      }
-
-   if(self()->comp()->getOption(TR_TraceCG))
-         traceMsg(self()->comp(), "\n");
-
-   return retval;
-   }
-
-
 ////////////////////////////////////////////////////////////////////////////////
 // TR::390SystemLinkage::createPrologue() - create prolog for System linkage
 ////////////////////////////////////////////////////////////////////////////////
-void TR::S390SystemLinkage::createPrologue(TR::Instruction * cursor)
+void TR::SystemLinkage::createPrologue(TR::Instruction * cursor)
    {
    TR::Delimiter delimiter (comp(), comp()->getOption(TR_TraceCG), "Prologue");
    TR::RealRegister * spReg = getNormalStackPointerRealRegister();
@@ -1711,7 +654,7 @@ void TR::S390SystemLinkage::createPrologue(TR::Instruction * cursor)
 
    //  We assume frame size is less than 32k
    //TR_ASSERT(stackFrameSize<=MAX_IMMEDIATE_VAL,
-   //      "TR::S390SystemLinkage::createPrologue -- Frame size (0x%x) greater than 0x7FFF\n",stackFrameSize);
+   //      "TR::SystemLinkage::createPrologue -- Frame size (0x%x) greater than 0x7FFF\n",stackFrameSize);
 
    //Now that stackFrame size is known, map stack
    mapStack(bodySymbol, stackFrameSize);
@@ -1822,7 +765,7 @@ void TR::S390SystemLinkage::createPrologue(TR::Instruction * cursor)
 
 
 TR::Instruction*
-TR::S390SystemLinkage::restoreGPRsInEpilogue2(TR::Instruction *cursor)
+TR::SystemLinkage::restoreGPRsInEpilogue2(TR::Instruction *cursor)
    {
    int16_t GPRSaveMask = 0;
    int32_t offset = 0;
@@ -1884,7 +827,7 @@ TR::S390SystemLinkage::restoreGPRsInEpilogue2(TR::Instruction *cursor)
    }
 
 TR::Instruction*
-TR::S390SystemLinkage::restoreGPRsInEpilogue(TR::Instruction *cursor)
+TR::SystemLinkage::restoreGPRsInEpilogue(TR::Instruction *cursor)
    {
    int16_t GPRSaveMask = 0;
    bool    hasRestoreSP = false;
@@ -1963,7 +906,7 @@ TR::S390SystemLinkage::restoreGPRsInEpilogue(TR::Instruction *cursor)
 /**
  * Create epilog for System linkage
  */
-void TR::S390SystemLinkage::createEpilogue(TR::Instruction * cursor)
+void TR::SystemLinkage::createEpilogue(TR::Instruction * cursor)
    {
    TR::Delimiter delimiter (comp(), comp()->getOption(TR_TraceCG), "Epilogue");
 
@@ -2024,921 +967,4 @@ void TR::S390SystemLinkage::createEpilogue(TR::Instruction * cursor)
    cursor = generateS390RegInstruction(cg(), TR::InstOpCode::BCR, nextNode,
           getRealRegister(REGNUM(TR::RealRegister::GPR14)), cursor);
    ((TR::S390RegInstruction *)cursor)->setBranchCondition(TR::InstOpCode::COND_BCR);
-   }
-
-int32_t
-TR::S390zLinuxSystemLinkage::getRegisterSaveOffset(TR::RealRegister::RegNum srcReg)
-   {
-   int32_t gpr2Offset = TR::Compiler->target.is64Bit() ? 16 : 8;
-   int32_t fpr0Offset = TR::Compiler->target.is64Bit() ? 128 : 64;
-   switch(srcReg)
-      {
-      case TR::RealRegister::FPR0: return fpr0Offset;
-      case TR::RealRegister::FPR2: return fpr0Offset + cg()->machine()->getFPRSize();
-      case TR::RealRegister::FPR4: return fpr0Offset + 2 * cg()->machine()->getFPRSize();
-      case TR::RealRegister::FPR6: return fpr0Offset + 3 * cg()->machine()->getFPRSize();
-      case TR::RealRegister::FPR1:
-      case TR::RealRegister::FPR3:
-      case TR::RealRegister::FPR5:
-      case TR::RealRegister::FPR7:
-      case TR::RealRegister::GPR0:
-      case TR::RealRegister::GPR1:
-         TR_ASSERT(false,"ERROR: TR::zLinuxSystemLinkage::getRegisterSaveOffset called for volatile reg: %d",srcReg);
-         return -1;
-      case TR::RealRegister::GPR2: return gpr2Offset;
-      default: return gpr2Offset + (srcReg - TR::RealRegister::GPR2) * cg()->machine()->getGPRSize();
-      }
-   }
-
-int32_t
-TR::S390zOSSystemLinkage::getRegisterSaveOffset(TR::RealRegister::RegNum srcReg)
-   {
-   int32_t offset;
-   if ((srcReg >= TR::RealRegister::GPR4) && (srcReg <= TR::RealRegister::GPR15))
-      {
-      int32_t offset = 2048 + (srcReg - TR::RealRegister::GPR4) * cg()->machine()->getGPRSize();
-      return offset;
-      }
-   else
-      {
-      TR_ASSERT(false, "ERROR: TR::S390zOSSystemLinkage::getRegisterSaveOffset called for volatile reg: %d\n",srcReg);
-      return -1;
-      }
-   }
-
-void OMR::Z::Linkage::replaceCallWithJumpInstruction(TR::Instruction *callInstruction)
-   {
-   TR::InstOpCode::Mnemonic opCode = callInstruction->getOpCodeValue();
-   TR_ASSERT(opCode == TR::InstOpCode::BASSM || opCode == TR::InstOpCode::BASR || opCode == TR::InstOpCode::BALR || opCode == TR::InstOpCode::BAS || opCode == TR::InstOpCode::BAL|| opCode == TR::InstOpCode::BRAS || opCode == TR::InstOpCode::BRASL, "Wrong opcode type on instruction!\n");
-
-   TR::Node *node = (callInstruction)->getNode();
-   TR::SymbolReference *callSymRef = (callInstruction)->getNode()->getSymbolReference();
-   TR::Symbol *callSymbol = (callInstruction)->getNode()->getSymbolReference()->getSymbol();
-
-   TR::Instruction *replacementInst =0 ;
-
-   replacementInst = new (self()->trHeapMemory()) TR::S390RILInstruction(TR::InstOpCode::BRCL, node, (uint32_t)0xf, callSymbol, callSymRef, self()->cg());
-
-   if(self()->comp()->getOption(TR_TraceCG))
-      traceMsg(self()->comp(), "Replacing instruction %p to a jump %p !\n",callInstruction, replacementInst);
-
-   self()->cg()->replaceInst(callInstruction,replacementInst);
-
-
-   }
-
-
-bool OMR::Z::Linkage::canDataTypeBePassedByReference(TR::DataType type) { return false; }
-
-bool TR::S390zLinuxSystemLinkage::canDataTypeBePassedByReference(TR::DataType type)
-   {
-
-   if (type == TR::Aggregate)
-      return true;
-
-   return false;
-   }
-
-bool OMR::Z::Linkage::isSymbolPassedByReference(TR::Symbol *sym) { return false; }
-
-bool TR::S390zLinuxSystemLinkage::isSymbolPassedByReference(TR::Symbol *sym)
-   {
-   if(!canDataTypeBePassedByReference(sym->getDataType()))
-      return false;
-
-   if(sym->getDataType() != TR::Aggregate)     //any non-aggregate symbols that have a data type that can be passed by reference *will* be passed by reference
-      return true;
-
-   int32_t gprSize = cg()->machine()->getGPRSize();
-   int32_t symSize = sym->getSize();
-   TR_ASSERT(0, "should not reach here due to folding of getDataTypeOfAggregateMember()");
-
-   TR::DataType aggrMemberDataType = TR::NoType;
-   bool isVectorAggregate = (TR::DataType(aggrMemberDataType).isVector() && cg()->getSupportsVectorRegisters());
-
-   if ((symSize > gprSize) && !(gprSize == 4 && symSize == 8) && !isVectorAggregate)
-      return true;
-
-   return false;
-   }
-
-bool
-TR::S390zOSSystemLinkage::isAggregateReturnedInIntRegistersAndMemory(int32_t aggregateLenth)
-   {
-   return isFastLinkLinkageType() &&
-      (aggregateLenth  > (getNumIntegerArgumentRegisters()* cg()->machine()->getGPRSize()));
-   }
-
-bool
-TR::S390zOSSystemLinkage::isAggregateReturnedInIntRegisters(int32_t aggregateLenth)
-   {
-   return aggregateLenth  <= (getNumIntegerArgumentRegisters()* cg()->machine()->getGPRSize());
-   }
-
-
-/**
- * General XPLink utility
- *
- * Calculate XPLink call descriptor for given call site (excl. entry offset).
- * "callNode"  can be null which means that this is a special call.
- * The retunred value is used for call descriptor emission and is also used to
- * determine floating point register likage for outgoing parameters.
- */
-uint32_t
-TR::S390zOSSystemLinkage::calculateCallDescriptorFlags(TR::Node *callNode)
-   {
-   #define XPLINK_FLOAT_PARM_UNPROTYPED_CALL 0x08
-   uint32_t callDescValue;
-
-   if (TR::Compiler->target.is64Bit())
-      {
-      return 0;
-      }
-
-   // Linkage field  (0 means XPLink)
-   // Bits 0-2 inclusive
-   //
-   bool linkageIsXPLink;
-   linkageIsXPLink = true;
-   callDescValue = 0x00000000;   // xplink (0) value here
-
-
-   //
-   // Return value adjust field
-   // Bits 3-7 inclusive
-   //
-
-   TR::DataType dataType;
-   TR::ILOpCodes opcode;
-   int32_t aggregateLength = 0;
-
-   if (callNode == NULL)  // specialized calls no real descriptor needed
-      dataType = TR::NoType;
-   else
-      {
-      dataType = callNode->getDataType();
-#ifdef J9_PROJECT_SPECIFIC
-      if (callNode->getType().isBCD())
-         aggregateLength = callNode->getSize();
-#endif
-      }
-
-   uint32_t rva;
-   if (linkageIsXPLink)
-      {
-      rva = calculateReturnValueAdjustFlag(dataType, aggregateLength);
-      }
-   else
-      {
-      // TOBEY generates 0 for non-XPLink cases.
-      // We are consistent with TOBEY although not sure that is 100% correct
-      rva = 0;
-      }
-   callDescValue |= rva << 24;
-
-   if (callNode == NULL)
-      {
-      return callDescValue; // specialized calls - so no float requirements
-      }
-
-   //
-   // Float parameter description fields
-   // Bits 8-31 inclusive
-   //
-
-   bool isPrototyped = true; // All Java calls are prototyped.
-   uint32_t parmAreaOffset = 0;
-
-#ifdef J9_PROJECT_SPECIFIC
-   TR::MethodSymbol* callSymbol = callNode->getSymbol()->castToMethodSymbol();
-   if (callSymbol->isJNI() && callNode->isPreparedForDirectJNI())
-      {
-      TR::ResolvedMethodSymbol * cs = callSymbol->castToResolvedMethodSymbol();
-      TR_ResolvedMethod * resolvedMethod = cs->getResolvedMethod();
-      // JNI Calls include a JNIEnv* pointer that is not included in list of children nodes.
-      // For FastJNI, certain calls do not require us to pass the JNIEnv.
-      if (!cg()->fej9()->jniDoNotPassThread(resolvedMethod))
-         parmAreaOffset += sizeof(uintptrj_t);
-
-      // For FastJNI, certain calls do not have to pass in receiver object.
-      if (cg()->fej9()->jniDoNotPassReceiver(resolvedMethod))
-         parmAreaOffset -= sizeof(uintptrj_t);
-      }
-#endif
-
-   uint32_t parmDescriptorFields = 0;
-
-   if (!isPrototyped)
-      {
-      parmDescriptorFields = shiftFloatParmDescriptorFlag(XPLINK_FLOAT_PARM_UNPROTYPED_CALL, 0);  // FPR0 field has special value for unprototped call
-      callDescValue |= parmDescriptorFields;
-      return callDescValue;
-      }
-
-   // WCode only logic follows for float parameter description fields
-
-   TR::Symbol *funcSymbol = callNode->getSymbolReference()->getSymbol();
-
-   uint32_t firstArgumentChild = callNode->getFirstArgumentIndex();
-   int32_t to = callNode->getNumChildren() - 1;
-   int32_t parmCount = 1;
-
-   int32_t floatParmNum = 0;
-
-   uint32_t lastFloatParmAreaOffset= 0;
-
-   bool done = false;
-   for (int32_t i = firstArgumentChild; (i <= to) && !done; i++, parmCount++)
-      {
-      TR::Node *child = callNode->getChild(i);
-      TR::DataType dataType = child->getDataType();
-      TR::SymbolReference *parmSymRef = child->getOpCode().hasSymbolReference() ? child->getSymbolReference() : NULL;
-      int32_t argSize = 0;
-
-      if (parmSymRef==NULL)
-         argSize = child->getSize();
-      else
-         argSize = parmSymRef->getSymbol()->getSize();
-
-      done = updateFloatParmDescriptorFlags(&parmDescriptorFields, funcSymbol, parmCount, argSize, dataType, &floatParmNum, &lastFloatParmAreaOffset, &parmAreaOffset);
-      }
-
-   callDescValue |= parmDescriptorFields;
-   return callDescValue;
-   }
-
-#define  GPREGINDEX(i)   (i-TR::RealRegister::FirstGPR)
-   
-// Calculate prologue info for XPLink
-// This is only done once for multiple entry point routines
-void TR::S390zOSSystemLinkage::calculatePrologueInfo(TR::Instruction * cursor)
-   {
-   TR::ResolvedMethodSymbol *bodySymbol = comp()->getJittedMethodSymbol();
-
-   // Logic for main most of which applies to secondary entry points
-   // I.E. all except for parameter offset information.
-
-   int32_t localSize, argSize;
-   int32_t stackFrameSize, alignedStackFrameSize;
-   int32_t stackFrameBias = getStackFrameBias();
-
-   // Set for initParamOffset() - relative offset from start of frame
-   setOutgoingParmAreaBeginOffset(getOffsetToFirstParm()-stackFrameBias);
-
-   {
-   //
-   // Calculate size of locals - determined in prior call to mapStack()
-   // We make the size a multiple of the strictest alignment symbol
-   // So that backwards mapping of auto symbols will follow the alignment
-   //
-   localSize =  -1 * (int32_t) (bodySymbol->getLocalMappingCursor());  // Auto+Spill size
-   int32_t strictestAutoAlign = 8;
-   localSize = (localSize + strictestAutoAlign - 1) & ~(strictestAutoAlign-1);
-
-   //
-   // Calculate size of outgoing argument area
-   //
-   argSize = getOutgoingParameterBlockSize();
-
-   //
-   // Calculate stack frame size
-   //
-   int32_t outGoingAreaOffsetUnbiased = (TR::Compiler->target.is64Bit()) ? 128 : 64;
-   stackFrameSize = outGoingAreaOffsetUnbiased + argSize + localSize;
-   int32_t minFrameAlign = (TR::Compiler->target.is64Bit()) ? 32 : 16;
-   TR_ASSERT( minFrameAlign >= strictestAutoAlign,"automatic symbol has alignment that cannot be satisfied");
-   alignedStackFrameSize = (stackFrameSize + minFrameAlign - 1) & ~(minFrameAlign-1);
-   stackFrameSize = alignedStackFrameSize;
-
-   setStackFrameSize(alignedStackFrameSize);
-
-   //
-   // Map stack - which is done in backwards fashion
-   //
-   int32_t stackMappingIndex = stackFrameBias + stackFrameSize;
-   mapStack(bodySymbol, stackMappingIndex);
-   }
-
-   setPrologueInfoCalculated(true);
-   }
-
-TR::Instruction * TR::S390zOSSystemLinkage::buyFrame(TR::Instruction * cursor, TR::Node * node)
-   {
-   enum TR_XPLinkFrameType frameType;
-   int16_t GPRSaveMask;
-   TR::LabelSymbol *stmLabel;
-
-   int8_t gprSize = cg()->machine()->getGPRSize();
-   int32_t stackFrameSize, offsetToRegSaveArea, offsetToFirstSavedReg, stackFrameBias, gpr3ParmOffset;
-   int32_t  firstSaved, lastSaved, firstPossibleSaved;
-   bool saveBackChain;
-
-   // TODO: enlarge this intermediate threshold for 64 bit (up to 1M). Needs changes in code below
-   int32_t intermediateThreshold = getGuardPageSize();
-
-   TR::RealRegister *spReg = getNormalStackPointerRealRegister(); // normal sp reg used in prol/epil
-   stackFrameSize = getStackFrameSize();
-   stackFrameBias = getStackFrameBias();
-   offsetToRegSaveArea = getOffsetToRegSaveArea();
-
-   // This delta is slightly pessimistic and could be a "tad" more. But, taking
-   // into account the "tad" complicates this code more than its worth.
-   // I.E. small set of cases could be forced to being intermediate vs. small
-   // frame types because of this "tad".
-   int32_t delta = stackFrameBias - stackFrameSize;
-
-   if (stackFrameSize > getGuardPageSize())
-      // guard page option mandates explicit checking
-      frameType = TR_XPLinkStackCheckFrame;
-   else if (delta >=0)
-      // delta > 0 implies stack offset for STM will be positive and <2K
-      frameType = TR_XPLinkSmallFrame;
-   else if (stackFrameSize < intermediateThreshold)
-      frameType = TR_XPLinkIntermediateFrame;
-   else
-      frameType = TR_XPLinkStackCheckFrame;
-   setFrameType(frameType);
-
-   TR::MemoryReference *rsa, *bosRef, *extenderRef, *gpr3ParmRef;
-   int32_t rsaOffset, bosOffset, extenderOffset;
-
-   //
-   // Determine minimal Preserved GPRs. Notes:
-   //   1) an option can force the save of the stack pointer register
-   //   2) XPLink noleaf:  routines mandates the save of the entry point register (r6) & return register (r7)
-   //               leaf:  routines cannot modify r6, r7
-   //
-   GPRSaveMask = 0;
-   TR::RealRegister::RegNum lastUsedReg = TR::RealRegister::NoReg;
-   for (int32_t i = TR::RealRegister::FirstGPR; i <= TR::RealRegister::LastGPR; ++i)
-     {
-     if (getPreserved(REGNUM(i)))
-        {
-        if (getRealRegister(REGNUM(i))->getHasBeenAssignedInMethod())
-           {
-           GPRSaveMask |= 1 << GPREGINDEX(i);
-           }
-        }
-     }
-
-   GPRSaveMask  |= 1 << GPREGINDEX(getReturnAddressRegister());
-   GPRSaveMask  |= 1 << GPREGINDEX(getEntryPointRegister());
-
-    // Determine if we need a temp register in either buying or releasing
-    // the stack frame. If so, we will save the back chain so epilogue code
-    // can just restore it from the frame and not require a temp register.
-    // We still need a temp register for prologue code though.
-    bool needAddTempReg, saveTempReg;
-    addImmediateToRealRegister(NULL, stackFrameSize, NULL, NULL, NULL, &needAddTempReg);
-
-   // For 64 bit stack checking code - need a temp reg for LAA access
-   if (TR::Compiler->target.is64Bit() && (frameType == TR_XPLinkStackCheckFrame))
-      needAddTempReg = true;
-
-   if (needAddTempReg)
-      setSaveBackChain(true);
-
-   saveBackChain = getSaveBackChain();
-   if (saveBackChain)
-      GPRSaveMask  |= 1 << GPREGINDEX(getNormalStackPointerRegister());
-   else
-      GPRSaveMask  &= ~(1 << GPREGINDEX(getNormalStackPointerRegister()));
-
-
-   // CAA is either locked or restored on RET - so remove it from mask
-   // But not in 64 bit code
-   if(!TR::Compiler->target.is64Bit())
-      {
-      GPRSaveMask  &= ~(1 << GPREGINDEX(getCAAPointerRegister()));
-      }
-   
-   if ((frameType == TR_XPLinkStackCheckFrame)
-      || ((frameType == TR_XPLinkIntermediateFrame) && saveBackChain))
-      { // GPR4 is saved but indirectly via GPR0 in stack check or medium case - so don't save GPR4 via STM
-      firstSaved = getFirstMaskedBit(GPRSaveMask&~(1 << GPREGINDEX(getNormalStackPointerRegister())));
-      }
-   else if (frameType == TR_XPLinkStackLeafFrame)
-      { // Were not decrementing stack pointer but need stack space. We need to generate
-        // STM into frame - starting with GPR4,5,6
-        // These are not added to restore mask
-      GPRSaveMask  |= 1 << GPREGINDEX(getEntryPointRegister());
-      GPRSaveMask  |= 1 << GPREGINDEX(getReturnAddressRegister());
-      firstSaved = getFirstMaskedBit(GPRSaveMask);
-      }
-   else
-      firstSaved = getFirstMaskedBit(GPRSaveMask);
-
-   setGPRSaveMask(GPRSaveMask);
-
-   if (comp()->getOption(TR_TraceCG))
-      {
-      traceMsg(comp(), "GPRSaveMask: Register context %x\n", GPRSaveMask&0xffff);
-      }
-
-   firstPossibleSaved = GPREGINDEX(TR::RealRegister::GPR4); // GPR4 is first reg in save area
-   lastSaved = getLastMaskedBit(GPRSaveMask);
-   offsetToFirstSavedReg = (firstSaved-firstPossibleSaved)*gprSize; // relative to start of save area
-
-   // variables for stack extension code - xplink spec mandates these particular registers
-   TR::RealRegister * gpr0Real = getRealRegister(TR::RealRegister::GPR0);
-   TR::RealRegister * gpr3Real = getRealRegister(TR::RealRegister::GPR3);
-   TR::RealRegister * caaReal  = getRealRegister(getCAAPointerRegister());  // 31 bit oly
-   gpr3ParmOffset = getOffsetToFirstParm() + 2 * gprSize;
-   
-   _firstSaved = REGNUM(firstSaved + TR::RealRegister::FirstGPR);
-   _lastSaved  = REGNUM(lastSaved  + TR::RealRegister::FirstGPR);
-
-   switch (frameType)
-      {
-      case TR_XPLinkNoStackLeafFrame:
-          break; // nothing to buy or save
-
-      case TR_XPLinkStackLeafFrame:
-          // STM rx,ry,offset(R4)
-          rsaOffset =  (offsetToRegSaveArea - stackFrameSize) + offsetToFirstSavedReg;
-          rsa = generateS390MemoryReference(spReg, rsaOffset, cg());
-          cursor = generateRSInstruction(cg(),  TR::InstOpCode::getStoreMultipleOpCode(),
-                  node, getRealRegister(REGNUM(firstSaved + TR::RealRegister::FirstGPR)),
-                  getRealRegister(REGNUM(lastSaved + TR::RealRegister::FirstGPR)), rsa, cursor);
-          break;  // not frame to buy but save regs
-
-      case TR_XPLinkSmallFrame:
-          // STM rx,ry,offset(R4)
-          rsaOffset =  (offsetToRegSaveArea - stackFrameSize) + offsetToFirstSavedReg;
-          rsa = generateS390MemoryReference(spReg, rsaOffset, cg());
-          cursor = generateRSInstruction(cg(),  TR::InstOpCode::getStoreMultipleOpCode(),
-                  node, getRealRegister(REGNUM(firstSaved + TR::RealRegister::FirstGPR)),
-                  getRealRegister(REGNUM(lastSaved + TR::RealRegister::FirstGPR)), rsa, cursor);
-
-          cursor = generateS390LabelInstruction(cg(), InstOpCode::LABEL, node, generateLabelSymbol(cg()), cursor);
-          _stackPointerUpdate = cursor;
-
-          // AHI R4,-stackFrameSize
-          cursor = addImmediateToRealRegister(spReg, (stackFrameSize) * -1, NULL, node, cursor);
-          break;
-
-      case TR_XPLinkIntermediateFrame:
-          if (needAddTempReg)
-             {
-             // (saveBackChain) LR R0,R4
-             // (saveTempReg|saveBackChain) ST R3,D(R4)   <-- save temp register
-             // R4<- R4 - stackSize
-             // LR R3,R0
-             // (saveTempReg) L  R3,D(R3)   <-- load temp register
-             saveTempReg = true;  // PERFORMANCE NOTE: TODO: set to false if GPR3 is not used as a parameter
-
-             if (saveTempReg || saveBackChain)
-                {
-                cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node,  gpr0Real, spReg, cursor); // LR R0,R4
-                }
-             if (saveTempReg)
-                {
-                gpr3ParmRef = generateS390MemoryReference(spReg, gpr3ParmOffset, cg()); // used for temporary register
-                cursor = generateRXInstruction(cg(), TR::InstOpCode::getStoreOpCode(), node, gpr3Real, gpr3ParmRef, cursor); // ST R3,parmRef(R4)
-                }
-
-             cursor = generateS390LabelInstruction(cg(), InstOpCode::LABEL, node, generateLabelSymbol(cg()), cursor);
-             _stackPointerUpdate = cursor;
-
-             cursor = addImmediateToRealRegister(spReg, (stackFrameSize) * -1, gpr3Real, node, cursor);  // R4 <- R4 - stack size
-
-             if (saveTempReg)
-                {
-                cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node,  gpr3Real, gpr0Real, cursor);
-                gpr3ParmRef = generateS390MemoryReference(gpr3Real, gpr3ParmOffset, cg());
-                cursor = generateRXInstruction(cg(), TR::InstOpCode::getLoadOpCode(), node, gpr3Real, gpr3ParmRef, cursor);
-                }
-             }
-          else
-             { // no temp reg needed
-             // (saveBackChain)    LR R0,R4
-             // AHI R4,-stackFrameSize
-             // STM rx,ry,offset(R4)
-             // (saveBackChain)    ST R0,D(R4)
-
-             if (saveBackChain) // LR R0,R4
-                cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node, gpr0Real, spReg, cursor);
-             
-             cursor = generateS390LabelInstruction(cg(), InstOpCode::LABEL, node, generateLabelSymbol(cg()), cursor);
-             _stackPointerUpdate = cursor;
-
-             cursor = addImmediateToRealRegister(spReg, (stackFrameSize) * -1, NULL, node, cursor);
-             }
-          rsaOffset =  offsetToRegSaveArea + offsetToFirstSavedReg;
-          rsa = generateS390MemoryReference(spReg, rsaOffset, cg());
-          cursor = generateRSInstruction(cg(),  TR::InstOpCode::getStoreMultipleOpCode(),
-                  node, getRealRegister(REGNUM(firstSaved + TR::RealRegister::FirstGPR)),
-                  getRealRegister(REGNUM(lastSaved + TR::RealRegister::FirstGPR)), rsa, cursor);
-          if (saveBackChain)
-             { // ST R0,2048(R4)
-             rsaOffset = offsetToRegSaveArea + 0; // GPR4 saved at start of RSA
-             rsa = generateS390MemoryReference(spReg, rsaOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getStoreOpCode(), node, gpr0Real, rsa, cursor);
-             }
-          break;
-
-      case TR_XPLinkStackCheckFrame:
-          // TODO: move the code to exend stack to "cold" path
-          // Code:
-          //   important note: saveTempReg -> saveBackChain
-          //
-          //     (saveTempReg)      ST R3,parmOffsetR3(GPR4)
-          //     (saveBackChain)    LR R0,R4
-          //     AHI R4, -stackFrameSize  [this varies based on stackFrameSize]
-          //     [31 bit] C  R4,BOS(R12)           R12=CAA
-          //     [64 bit] LLGT  R3,1268            R3=control block - low mem
-          //     [64 bit] C  R4,BOS(R3)            R3=ditto
-          //     BL stmLabel
-          //     /* extension logic */
-          //     (!saveBackChain)    LR R0,R3   /* could be smarter here */
-          //     L  R3,extender(R12)
-          //     BASR R3,R3
-          //     NOP x                    [xplink noop]
-          //     (!saveBackChain)   LR R3,R0   /* could be smarter here */
-          //
-          //   stmLabel:
-          //     STM rx,ry,offset(R4)
-          //     (saveBackChain)    ST R0,2048(,R4)
-          //     (saveTempReg)      LR R3,R0   /* could be smarter here */
-          //     (saveTempReg)      L  R3,parmOffsetR3(GPR3)
-          //
-          stmLabel = generateLabelSymbol(cg());
-          saveTempReg = saveBackChain;  // TODO: and with check that R3 is used as parm
-
-          //------------------------------
-          // Stack check - prologue part 1
-          //------------------------------
-
-          if (saveTempReg)
-             {
-             gpr3ParmRef = generateS390MemoryReference(spReg, gpr3ParmOffset, cg()); // used for temporary register
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getStoreOpCode(), node, gpr3Real, gpr3ParmRef, cursor); // ST R3,parmRef(R4)
-             }
-          if (saveBackChain) // LR R0,R4
-              cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node,  gpr0Real, spReg, cursor);
-          cursor = addImmediateToRealRegister(spReg, (stackFrameSize) * -1, gpr3Real, node, cursor); // AHI ...
-
-          if (TR::Compiler->target.is64Bit())
-             {
-             TR::MemoryReference *laaRef;
-             bosOffset = 64; // 64=offset in LAA to BOS
-             int32_t laaLowMemAddress = 1208; // LAA address
-             laaRef = generateS390MemoryReference(gpr0Real, laaLowMemAddress, cg());
-             cursor = generateRXInstruction(cg(), InstOpCode::LLGT, node, gpr3Real, laaRef, cursor);
-             bosRef = generateS390MemoryReference(gpr3Real, bosOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getCmpOpCode(), node, spReg, bosRef, cursor); // C R4,bos(tempreg)
-             }
-          else
-             {
-             bosOffset = 868; // 868=offset in CAA to BOS
-             bosRef = generateS390MemoryReference(caaReal, bosOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getCmpOpCode(), node, spReg, bosRef, cursor); // C R4,bos(R12)
-             }
-          cursor = generateS390BranchInstruction(cg(), InstOpCode::BRC, InstOpCode::COND_BNM, node, stmLabel, cursor); // BNM stmLabel
-
-          //------------------------------
-          // Stack check - extension routine logic
-          //------------------------------
-
-          // This logic should be moved out of line eventually
-          if (!saveBackChain)
-             {
-             cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node,  gpr0Real, gpr3Real, cursor); // LR R0,R3
-             }
-
-          if (TR::Compiler->target.is64Bit())
-             {
-             extenderOffset = 72;  // offset of extender routine in LAA
-             extenderRef = generateS390MemoryReference(gpr3Real, extenderOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getLoadOpCode(), node, gpr3Real, extenderRef, cursor); //  L R3,extender(R12)
-             }
-          else
-             {
-             extenderOffset = 872;  // offset of extender routine in CAA
-             extenderRef = generateS390MemoryReference(caaReal, extenderOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getLoadOpCode(), node, gpr3Real, extenderRef, cursor); //  L R3,extender(R12)
-             }
-          cursor = generateRRInstruction(cg(), InstOpCode::BASR, node, gpr3Real, gpr3Real, cursor);
-          cursor = genCallNOPAndDescriptor(cursor, node, NULL, TR_XPLinkCallType_BASR33);
-
-          if (!saveBackChain)
-             { // LR R3,R0
-             cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node,  gpr3Real, gpr0Real, cursor);
-             }
-
-          //------------------------------
-          // Stack check - prologue part 2
-          //------------------------------
-
-          // STM rx,ry,disp(R4)
-          cursor = generateS390LabelInstruction(cg(), InstOpCode::LABEL, node, stmLabel, cursor);
-
-          rsaOffset =  offsetToRegSaveArea + offsetToFirstSavedReg;
-          rsa = generateS390MemoryReference(spReg, rsaOffset, cg());
-          cursor = generateRSInstruction(cg(),  TR::InstOpCode::getStoreMultipleOpCode(),
-                  node, getRealRegister(REGNUM(firstSaved + TR::RealRegister::FirstGPR)),
-                  getRealRegister(REGNUM(lastSaved + TR::RealRegister::FirstGPR)), rsa, cursor);
-          if (saveBackChain)
-             { // ST R0,2048(R4)
-             rsaOffset = offsetToRegSaveArea + 0; // GPR4 saved at start of RSA
-             rsa = generateS390MemoryReference(spReg, rsaOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getStoreOpCode(), node, gpr0Real, rsa, cursor);
-             }
-          if (saveTempReg)
-             {
-             cursor = generateRRInstruction(cg(), TR::InstOpCode::getLoadRegOpCode(), node,  gpr3Real, gpr0Real, cursor);
-             gpr3ParmRef = generateS390MemoryReference(gpr3Real, gpr3ParmOffset, cg());
-             cursor = generateRXInstruction(cg(), TR::InstOpCode::getLoadOpCode(), node, gpr3Real, gpr3ParmRef, cursor);
-             }
-
-          break;
-
-      default:
-          TR_ASSERT( 0,"invalid xplink frame type");
-          break;
-      }
-
-   return cursor;
-   }
-
-void
-TR::S390zOSSystemLinkage::createEntryPointMarker(TR::Instruction* cursor, TR::Node* node)
-   {
-   }
-
-TR_XPLinkCallTypes
-TR::S390zOSSystemLinkage::genWCodeCallBranchCode(TR::Node *callNode, TR::RegisterDependencyConditions * deps)
-   {
-   // WCode specific
-   //
-   // There are 4 cases for outgoing branch sequences
-   //   case 1)            pure OS linkage call using entry point
-   //                      BASR R14,R15
-   //   case 2) (indirect) Call vi function pointer
-   //                      LM R5,R6,disp(regfp)
-   //                      BASR R7,R6
-   //          where:
-   //             a) regfp is a register containing pointer to function descriptor
-   //                (most likely placed in R6)
-   //             b) disp is offset into function descriptor - 16/0 (31/64 bit respectively)
-   //   case 3) (direct) Call to (static or global) function defined in the compilation unit:
-   //                      BRASL R7,func
-   //
-   //   case 4) (direct) Call to external function referenced function
-   //                      LM R5,R6,disp(regenv)
-   //                      BASR R7,R6
-   //         where:
-   //             a) disp is an offset in the environment (aka ADA) containing the
-   //                function descriptor body (i.e. not pointer to function descriptor)
-   TR_XPLinkCallTypes callType;
-
-   // Find GPR6 (xplink) or GPR15 (os linkage) in post conditions
-   TR::Register * systemEntryPointRegister = deps->searchPostConditionRegister(getEntryPointRegister());
-   // Find GPR7 (xplik) or GPR14 (os linkage) in post conditions
-   TR::Register * systemReturnAddressRegister = deps->searchPostConditionRegister(getReturnAddressRegister());
-
-   TR::RegisterDependencyConditions * preDeps = new (trHeapMemory()) TR::RegisterDependencyConditions(deps->getPreConditions(), NULL, deps->getAddCursorForPre(), 0, cg());
-
-   if (callNode->getOpCode().isIndirect())
-      {
-      generateRRInstruction(cg(), InstOpCode::BASR, callNode, systemReturnAddressRegister, systemEntryPointRegister, preDeps);
-      callType = TR_XPLinkCallType_BASR;
-      return callType;
-      }
-
-   TR::SymbolReference *callSymRef = callNode->getSymbolReference();
-   TR::Symbol *callSymbol = callSymRef->getSymbol();
-
-   TR::Instruction * callInstr = new (cg()->trHeapMemory()) TR::S390RILInstruction(TR::InstOpCode::BRASL, callNode, systemReturnAddressRegister, callSymbol, callSymRef, cg());
-   callInstr->setDependencyConditions(preDeps);
-   callType = TR_XPLinkCallType_BRASL7;
-
-   return callType;
-   }
-
-/////////////////////////////////////////////////////////////////////////////////
-// TR::S390zOSSystemLinkage::generateInstructionsForCall - Front-end
-//  customization of callNativeFunction
-////////////////////////////////////////////////////////////////////////////////
-void
-TR::S390zOSSystemLinkage::generateInstructionsForCall(TR::Node * callNode, TR::RegisterDependencyConditions * deps, intptrj_t targetAddress,
-      TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
-      TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint)
-   {
-   TR::CodeGenerator * codeGen = cg();
-
-    TR_XPLinkCallTypes callType;  // for XPLink calls
-
-    // only pre-dependencies from dep will be attached to the call
-    callType = genWCodeCallBranchCode(callNode, deps);
-
-    generateS390LabelInstruction(codeGen, InstOpCode::LABEL, callNode, returnFromJNICallLabel);
-
-    genCallNOPAndDescriptor(NULL, callNode, callNode, callType);
-
-    // Append post-dependencies after NOP
-    TR::LabelSymbol * afterNOP = generateLabelSymbol(cg());
-    TR::RegisterDependencyConditions * postDeps =
-          new (trHeapMemory()) TR::RegisterDependencyConditions(NULL,
-                deps->getPostConditions(), 0, deps->getAddCursorForPost(), cg());
-    generateS390LabelInstruction(codeGen, InstOpCode::LABEL, callNode, afterNOP, postDeps);
-}
-
-TR::Instruction*
-TR::S390zOSSystemLinkage::getStackPointerUpdate() const
-   {
-   return _stackPointerUpdate;
-   }
-
-TR::PPA1Snippet*
-TR::S390zOSSystemLinkage::getPPA1Snippet() const
-   {
-   return _ppa1;
-   }
-
-TR::PPA2Snippet*
-TR::S390zOSSystemLinkage::getPPA2Snippet() const
-   {
-   return _ppa2;
-   }
-
-// Create prologue for XPLink
-void TR::S390zOSSystemLinkage::createPrologue(TR::Instruction * cursor)
-   {
-   TR_ASSERT( cursor!=NULL,"cursor is NULL in createPrologue");
-   calculatePrologueInfo(cursor);
-
-   TR::ResolvedMethodSymbol *bodySymbol = comp()->getJittedMethodSymbol();
-
-   TR::Node *node = cursor->getNode();
-
-   TR::CodeGenerator* cg = self()->cg();
-
-   _ppa1 = new (self()->trHeapMemory()) TR::PPA1Snippet(cg, this);
-   _ppa2 = new (self()->trHeapMemory()) TR::PPA2Snippet(cg, this);
-
-   cg->addSnippet(_ppa1);
-   cg->addSnippet(_ppa2);
-
-   uint32_t stackFrameSize = getStackFrameSize();
-
-   TR_ASSERT_FATAL((stackFrameSize & 31) == 0, "XPLINK stack frame size (%d) has to be aligned to 32-bytes.", stackFrameSize);
-   
-   cursor = cursor->getPrev();
-
-   // "C.E.E.1"
-   cursor = generateDataConstantInstruction(cg, TR::InstOpCode::DC, node, 0x00C300C5, cursor);
-   cursor = generateDataConstantInstruction(cg, TR::InstOpCode::DC, node, 0x00C500F1, cursor);
-   cursor = generateDataConstantInstruction(cg, TR::InstOpCode::DC, node, 0x00000000, cursor);
-
-   cg->addRelocation(new (cg->trHeapMemory()) EntryPointMarkerOffsetToPPA1Relocation(cursor, _ppa1->getSnippetLabel()));
-   
-   // DSA size is the frame size aligned to 32-bytes which means it's least significant 5 bits are zero and are used to
-   // represent the flags which are always 0 for OMR as we do not support leaf frames or direct calls to alloca()
-   cursor = generateDataConstantInstruction(cg, TR::InstOpCode::DC, node, stackFrameSize, cursor);
-
-   cursor = cursor->getNext();
-
-   setFirstPrologueInstruction(cursor);
-
-   //
-   // Generate code to acquire the stack frame
-   //
-   cursor = buyFrame(cursor, node);
-
-   //
-   // Save FPRs and ARs in local automatic
-   // This is done after mapStack - as it calculates the register mask and
-   // save area offset.
-   //
-   cursor = getputFPRs(InstOpCode::STD, cursor, node);
-
-   //
-   // Save arguments as required
-   //
-   cursor = (TR::Instruction *) saveArguments(cursor, false);
-   
-   // Last prologue instruction
-   setLastPrologueInstruction(cursor);
-   }
-
-void
-TR::S390zOSSystemLinkage::createEpilogue(TR::Instruction * cursor)
-   {
-   int16_t GPRSaveMask;
-   int8_t gprSize = cg()->machine()->getGPRSize();
-   enum TR_XPLinkFrameType frameType = getFrameType();
-   TR::Node * currentNode = cursor->getNode();
-   int32_t offset;
-
-   int32_t stackFrameSize, offsetToFirstSavedReg, stackFrameBias;
-   int32_t  firstSaved, lastSaved, firstPossibleSaved;
-   TR::MemoryReference *rsa;
-
-   int32_t blockNumber = -1;
-
-   TR::RealRegister *spReg = getNormalStackPointerRealRegister(); // normal sp reg used in prol/epil
-   stackFrameSize = getStackFrameSize();
-   stackFrameBias = getStackFrameBias();
-
-   //
-   // Restore FPRs and ARs from local automatic
-   //
-   cursor = getputFPRs(InstOpCode::LD, cursor, currentNode);
-
-   GPRSaveMask = getGPRSaveMask(); // restore mask is subset of save mask
-   GPRSaveMask  &= ~(1 << GPREGINDEX(getEntryPointRegister())); // entry point register not restored
-
-   firstSaved = getFirstMaskedBit(GPRSaveMask);
-   firstPossibleSaved = 4; // GPR4 is first reg in save area
-   lastSaved = getLastMaskedBit(GPRSaveMask);
-
-   if (firstSaved >= 0)
-      {
-      offsetToFirstSavedReg = (firstSaved-firstPossibleSaved)*gprSize; // relative to start of save area
-      offset = stackFrameBias + offsetToFirstSavedReg;
-      rsa = generateS390MemoryReference(spReg, offset, cg());
-
-      if (firstSaved == lastSaved)
-         {
-            cursor = generateRXInstruction(cg(), TR::InstOpCode::getLoadOpCode(), currentNode,
-                                           getRealRegister(REGNUM(firstSaved + TR::RealRegister::FirstGPR)), rsa, cursor);
-         }
-      else if (firstSaved < lastSaved)
-         {
-         int8_t numNeededToRestore = -1;
-
-            cursor = restorePreservedRegs(REGNUM(firstSaved + TR::RealRegister::FirstGPR),
-                                         REGNUM(lastSaved + TR::RealRegister::FirstGPR),
-                                         blockNumber, cursor, currentNode, spReg, rsa, getNormalStackPointerRegister());
-         }
-      }
-
-   if (!(GPRSaveMask &(1 << GPREGINDEX(getNormalStackPointerRegister()))))
-      { // GPR4 not restored by previous and needs restoring
-      cursor = addImmediateToRealRegister(spReg, stackFrameSize,  NULL, currentNode, cursor);
-      TR_ASSERT( cursor != NULL, "xplink retore code - should not need temp register");
-      }
-
-   // 31 bit - B 4(,7),  64 bit B 2(,7)
-   TR::RealRegister *dummyUncondMask = getRealRegister(REGNUM(TR::RealRegister::GPR15));
-   TR::RealRegister * realReturnRegister = getReturnAddressRealRegister();
-   offset = (TR::Compiler->target.is64Bit()) ? 2 : 4;
-   cursor = generateRXInstruction(cg(), InstOpCode::BC, currentNode, dummyUncondMask, generateS390MemoryReference(realReturnRegister, offset, cg()), cursor);
-   }
-
-/**
- * General XPLink utility
- */
-TR::Instruction *
-TR::S390zOSSystemLinkage::genCallNOPAndDescriptor(TR::Instruction * cursor,
-                                                    TR::Node *node,
-                                                    TR::Node *callNode,
-                                                    TR_XPLinkCallTypes callType)
-   {
-   TR::CodeGenerator * codeGen = cg();
-   // In 64 bit XPLINK, the caller returns at RetAddr+2, so add 2 byte nop
-   // In 31 bit XPLINK, the caller returns at RetAddr+4, so add 4 byte nop with last two bytes
-   // as signed offset in doublewords at or preceding NOP
-   uint32_t padSize = TR::Compiler->target.is64Bit() ? 2 : 4;
-
-   cursor = codeGen->insertPad(node, cursor, padSize, false);
-   TR::S390NOPInstruction *nop = static_cast<TR::S390NOPInstruction *>(cursor);
-
-   if (TR::Compiler->target.is32Bit())
-      {
-      uint32_t callDescValues = calculateCallDescriptorFlags(callNode);  // lower 32 bits
-
-      TR::S390ConstantDataSnippet * callDescSnippet = cg()->findOrCreate8ByteConstant(node, (int64_t)callDescValues, false);
-      nop->setTargetSnippet(callDescSnippet);
-
-      // Create a pseudo instruction that will generate
-      //      BRAS 4
-      //      DC <Call Descriptor>
-      // if the snippet is > 15k bytes away.
-
-      TR::S390PseudoInstruction *pseudoCallDesc = static_cast<TR::S390PseudoInstruction *>(generateS390PseudoInstruction(codeGen, TR::InstOpCode::XPCALLDESC, node, NULL, cursor));
-      // Assign this pseudoCallDesc to the NOP Instruction.
-      static_cast<TR::S390NOPInstruction *>(cursor)->setCallDescInstr(pseudoCallDesc);
-      cursor = pseudoCallDesc;
-      }
-
-   nop->setCallType(callType);
-   return cursor;
-   }
-
-TR::S390zOSSystemLinkage::EntryPointMarkerOffsetToPPA1Relocation::EntryPointMarkerOffsetToPPA1Relocation(TR::Instruction* cursor, TR::LabelSymbol* ppa2)
-   :
-      TR::LabelRelocation(NULL, ppa2),
-      _cursor(cursor)
-   {
-   }
-
-void
-TR::S390zOSSystemLinkage::EntryPointMarkerOffsetToPPA1Relocation::apply(TR::CodeGenerator* cg)
-   {
-   uint8_t* cursor = _cursor->getBinaryEncoding();
-
-   // The 0x08 is the static distance from the field which we need to relocate to the start of the Entry Point Marker
-   *reinterpret_cast<int32_t*>(cursor) = static_cast<int32_t>(getLabel()->getCodeLocation() - (cursor - 0x08));
    }
