+++ compiler/z/codegen/SystemLinkage.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2016 IBM Corp. and others
+ * Copyright (c) 2000, 2019 IBM Corp. and others
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which accompanies this
@@ -22,22 +22,188 @@
 #ifndef TR_SYSTEMLINKAGE_INCL
 #define TR_SYSTEMLINKAGE_INCL
 
-#include "codegen/S390SystemLinkage.hpp"
-
+#include <stddef.h>
+#include <stdint.h>
+#include "codegen/InstOpCode.hpp"
 #include "codegen/Linkage.hpp"
 #include "codegen/LinkageConventionsEnum.hpp"
+#include "codegen/RealRegister.hpp"
+#include "codegen/Register.hpp"
+#include "compile/Compilation.hpp"
+#include "env/TRMemory.hpp"
+#include "env/jittypes.h"
+#include "il/Symbol.hpp"
+#include "il/SymbolReference.hpp"
+#include "il/symbol/AutomaticSymbol.hpp"
+#include "infra/Array.hpp"
+#include "infra/Assert.hpp"
 
+namespace TR { class S390JNICallDataSnippet; }
+namespace TR { class Block; }
 namespace TR { class CodeGenerator; }
+namespace TR { class Instruction; }
+namespace TR { class LabelSymbol; }
+namespace TR { class Node; }
+namespace TR { class ParameterSymbol; }
+namespace TR { class RegisterDependencyConditions; }
+namespace TR { class ResolvedMethodSymbol; }
+namespace TR { class SystemLinkage; }
+template <class T> class List;
 
 namespace TR
 {
 
-class SystemLinkage : public OMR::SystemLinkageConnector
+class SystemLinkage : public TR::Linkage
    {
    public:
 
-   SystemLinkage(TR::CodeGenerator * cg, TR_S390LinkageConventions elc=TR_S390LinkageDefault, TR_LinkageConventions lc=TR_System)
-      : OMR::SystemLinkageConnector(cg, elc, lc) {}
+   SystemLinkage(TR::CodeGenerator * cg, TR_S390LinkageConventions elc = TR_S390LinkageDefault, TR_LinkageConventions lc = TR_System);
+
+   TR::SystemLinkage * self();
+
+   virtual int32_t getRegisterSaveOffset(TR::RealRegister::RegNum) = 0;
+
+   int32_t setStackFrameSize(int32_t StackFrameSize)  { return _StackFrameSize = StackFrameSize; }
+   virtual int32_t getStackFrameSize()                { return _StackFrameSize; }
+
+   int16_t setGPRSaveMask(int16_t GPRSaveMask)  { return _GPRSaveMask = GPRSaveMask; }
+   int16_t getGPRSaveMask()                     { return _GPRSaveMask; }
+
+   int16_t setFPRSaveMask(int16_t FPRSaveMask)  { return _FPRSaveMask = FPRSaveMask; }
+   int16_t getFPRSaveMask()                     { return _FPRSaveMask; }
+
+   static uint16_t flipBitsRegisterSaveMask(uint16_t mask);
+
+   int32_t setGPRSaveAreaBeginOffset(int32_t GPRSaveAreaBeginOffset)  { return _GPRSaveAreaBeginOffset = GPRSaveAreaBeginOffset; }
+   int32_t getGPRSaveAreaBeginOffset()                                { return _GPRSaveAreaBeginOffset; }
+   int32_t setGPRSaveAreaEndOffset(int32_t GPRSaveAreaEndOffset)  { return _GPRSaveAreaEndOffset = GPRSaveAreaEndOffset; }
+   int32_t getGPRSaveAreaEndOffset()                              { return _GPRSaveAreaEndOffset; }
+
+   int32_t setFPRSaveAreaBeginOffset(int32_t FPRSaveAreaBeginOffset)  { return _FPRSaveAreaBeginOffset = FPRSaveAreaBeginOffset; }
+   int32_t getFPRSaveAreaBeginOffset()                                { return _FPRSaveAreaBeginOffset; }
+   int32_t setFPRSaveAreaEndOffset(int32_t FPRSaveAreaEndOffset)  { return _FPRSaveAreaEndOffset = FPRSaveAreaEndOffset; }
+   int32_t getFPRSaveAreaEndOffset()                              { return _FPRSaveAreaEndOffset; }
+
+   int32_t setOutgoingParmAreaBeginOffset(int32_t OutgoingParmAreaBeginOffset)    { return _OutgoingParmAreaBeginOffset = OutgoingParmAreaBeginOffset; }
+   int32_t getOutgoingParmAreaBeginOffset()                                 { return _OutgoingParmAreaBeginOffset; }
+   int32_t setOutgoingParmAreaEndOffset(int32_t OutgoingParmAreaEndOffset) { return _OutgoingParmAreaEndOffset = OutgoingParmAreaEndOffset; }
+   int32_t getOutgoingParmAreaEndOffset()                                  { return _OutgoingParmAreaEndOffset; }
+
+   virtual int32_t setLocalsAreaBeginOffset(int32_t LocalsAreaBeginOffset)    { return _LocalsAreaBeginOffset = LocalsAreaBeginOffset; }
+   virtual int32_t getLocalsAreaBeginOffset()                                 { return _LocalsAreaBeginOffset; }
+   virtual int32_t setLocalsAreaEndOffset(int32_t LocalsAreaEndOffset)        { return _LocalsAreaEndOffset = LocalsAreaEndOffset; }
+   virtual int32_t getLocalsAreaEndOffset()                                   { return _LocalsAreaEndOffset; }
+
+   int32_t setNumUsedArgumentGPRs(int32_t numUsedArgumentGPRs)    { return _numUsedArgumentGPRs = numUsedArgumentGPRs; }
+   int32_t getNumUsedArgumentGPRs()  { return _numUsedArgumentGPRs; }
+
+   int32_t setNumUsedArgumentFPRs(int32_t numUsedArgumentFPRs)    { return _numUsedArgumentFPRs = numUsedArgumentFPRs; }
+   int32_t getNumUsedArgumentFPRs()  { return _numUsedArgumentFPRs; }
+
+   int32_t setNumUsedArgumentVFRs(int32_t numUsedArgumentVFRs)    { return _numUsedArgumentVFRs = numUsedArgumentVFRs; }
+   int32_t getNumUsedArgumentVFRs()  { return _numUsedArgumentVFRs; }
+
+   int32_t setVarArgOffsetInParmArea(int32_t varArgOffsetInParmArea)    { return _varArgOffsetInParmArea = varArgOffsetInParmArea; }
+   int32_t getVarArgOffsetInParmArea()  { return _varArgOffsetInParmArea; }
+
+   int32_t setVarArgRegSaveAreaOffset(int32_t varArgRegSaveAreaOffset)    { return _varArgRegSaveAreaOffset = varArgRegSaveAreaOffset; }
+   int32_t getVarArgRegSaveAreaOffset()  { return _varArgRegSaveAreaOffset; }
+
+   void setIncomingParmAreaBeginOffset(int32_t offset) { _incomingParmAreaBeginOffset = offset; }
+   int32_t getIncomingParmAreaBeginOffset() { return _incomingParmAreaBeginOffset; }
+   void setIncomingParmAreaEndOffset(int32_t offset) { _incomingParmAreaEndOffset = offset; }
+   int32_t getIncomingParmAreaEndOffset() { return _incomingParmAreaEndOffset; }
+   virtual int32_t getIncomingParameterBlockSize() { return _incomingParmAreaEndOffset - _incomingParmAreaBeginOffset; }
+   virtual int32_t setParmOffsetInLocalArea(int32_t ParmOffsetInLocalArea)   { return _parmOffsetInLocalArea = ParmOffsetInLocalArea; }
+   virtual int32_t getParmOffsetInLocalArea()              { return _parmOffsetInLocalArea; }
+
+   virtual uint32_t
+   getIntArgOffset(int32_t index)
+      {
+      return 0;
+      }
+
+   virtual void generateInstructionsForCall(TR::Node * callNode, TR::RegisterDependencyConditions * dependencies,
+         intptrj_t targetAddress, TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel,
+         TR::S390JNICallDataSnippet * jniCallDataSnippet, bool isJNIGCPoint = true);
+
+   virtual TR::Register * callNativeFunction(TR::Node * callNode, TR::RegisterDependencyConditions * dependencies,
+      intptrj_t targetAddress, TR::Register * methodAddressReg, TR::Register * javaLitOffsetReg, TR::LabelSymbol * returnFromJNICallLabel, TR::S390JNICallDataSnippet * jniCallDataSnippet,
+      bool isJNIGCPoint = true);
+
+   virtual TR::Register *
+   buildDirectDispatch(TR::Node * callNode)
+      {
+      return buildSystemLinkageDispatch(callNode);
+      }
+
+   virtual TR::Register *
+   buildIndirectDispatch(TR::Node * callNode)
+      {
+      return buildSystemLinkageDispatch(callNode);
+      }
+
+   virtual void mapStack(TR::ResolvedMethodSymbol * symbol);
+   virtual void mapStack(TR::ResolvedMethodSymbol * method, uint32_t stackIndex);
+   virtual void mapSingleAutomatic(TR::AutomaticSymbol * p, uint32_t & stackIndex);
+   virtual bool hasToBeOnStack(TR::ParameterSymbol * parm);
+
+   virtual void createPrologue(TR::Instruction * cursor);
+   virtual void createEpilogue(TR::Instruction * cursor);
+
+   virtual TR::Instruction* saveGPRsInPrologue(TR::Instruction * cursor);
+   virtual TR::Instruction* saveGPRsInPrologue2(TR::Instruction * cursor);
+   virtual TR::Instruction* restoreGPRsInEpilogue(TR::Instruction *cursor);
+   virtual TR::Instruction* restoreGPRsInEpilogue2(TR::Instruction *cursor);
+
+   virtual FrameType checkLeafRoutine(int32_t stackFrameSize, TR::Instruction **callInstruction = 0) { return standardFrame; }      // default implementation.  Should be overridden
+   virtual FrameType checkLeafRoutine(int32_t stackFrameSize) { return standardFrame; }      // default implementation.  Should be overridden
+   virtual void initS390RealRegisterLinkage();
+
+   virtual TR::RealRegister::RegNum setNormalStackPointerRegister  (TR::RealRegister::RegNum r) { return _normalStackPointerRegister = r; }
+   virtual TR::RealRegister::RegNum getNormalStackPointerRegister()   { return _normalStackPointerRegister; }
+   virtual TR::RealRegister *getNormalStackPointerRealRegister() {return getRealRegister(_normalStackPointerRegister);}
+
+   virtual TR::RealRegister::RegNum setAlternateStackPointerRegister  (TR::RealRegister::RegNum r) { return _alternateStackPointerRegister = r; }
+   virtual TR::RealRegister::RegNum getAlternateStackPointerRegister()   { return _alternateStackPointerRegister; }
+   virtual TR::RealRegister *getAlternateStackPointerRealRegister() {return getRealRegister(_alternateStackPointerRegister);}
+   virtual void initParamOffset(TR::ResolvedMethodSymbol * method, int32_t stackIndex, List<TR::ParameterSymbol> *parameterList=0);
+
+   // Register used for EX instructions for debug hooks - set by notifyHasHooks
+   virtual TR::RealRegister::RegNum setDebugHooksRegister(TR::RealRegister::RegNum r) { return _debugHooksRegister = r; }
+   virtual TR::RealRegister::RegNum getDebugHooksRegister()   { return _debugHooksRegister; }
+   virtual TR::RealRegister *getDebugHooksRealRegister() {return getRealRegister(_debugHooksRegister);}
+
+   // == General utilities (linkage independent)
+   virtual TR::Instruction *addImmediateToRealRegister(TR::RealRegister * targetReg, int32_t immediate, TR::RealRegister *tempReg, TR::Node *node, TR::Instruction *cursor, bool *checkTempNeeded=NULL);
+   virtual TR::Instruction *getputFPRs(TR::InstOpCode::Mnemonic opcode, TR::Instruction *cursor, TR::Node *node, TR::RealRegister *spReg=0);
+   
+   protected:
+   int32_t _StackFrameSize;
+
+   private:
+
+   TR::RealRegister::RegNum _normalStackPointerRegister;
+   TR::RealRegister::RegNum _alternateStackPointerRegister;
+   TR::RealRegister::RegNum _debugHooksRegister;
+   int16_t _GPRSaveMask;
+   int16_t _FPRSaveMask;
+   int32_t _incomingParmAreaBeginOffset;
+   int32_t _incomingParmAreaEndOffset;
+   int32_t _FPRSaveAreaBeginOffset;
+   int32_t _FPRSaveAreaEndOffset;
+   int32_t _LocalsAreaBeginOffset;
+   int32_t _LocalsAreaEndOffset;
+   int32_t _OutgoingParmAreaBeginOffset;
+   int32_t _OutgoingParmAreaEndOffset;
+   int32_t _GPRSaveAreaBeginOffset;
+   int32_t _GPRSaveAreaEndOffset;
+   int32_t _numUsedArgumentGPRs;
+   int32_t _numUsedArgumentFPRs;
+   int32_t _numUsedArgumentVFRs;
+   int32_t _varArgOffsetInParmArea;
+   int32_t _varArgRegSaveAreaOffset;
+   int32_t _parmOffsetInLocalArea;
    };
 }
