+++ compiler/z/codegen/snippet/PPA2Snippet.cpp
@@ -19,7 +19,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0 OR GPL-2.0 WITH Classpath-exception-2.0 OR LicenseRef-GPL-2.0 WITH Assembly-exception
  *******************************************************************************/
 
-#include "codegen/TRSystemLinkage.hpp"
+#include "codegen/SystemLinkagezOS.hpp"
 #include "codegen/snippet/PPA2Snippet.hpp"
 
 TR::PPA2Snippet::PPA2Snippet(TR::CodeGenerator* cg, TR::S390zOSSystemLinkage* linkage)
