+++ fvtest/compilertest/z/codegen/TestCodeGenerator.cpp
@@ -20,7 +20,6 @@
  *******************************************************************************/
 
 #include "codegen/CodeGenerator.hpp"
-#include "codegen/TRSystemLinkage.hpp"
 #include "compile/Compilation.hpp"
 #include "env/CompilerEnv.hpp"
